% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-class.R
\name{add_class}
\alias{add_class}
\title{Add a Class to an Object}
\usage{
add_class(x, class)
}
\arguments{
\item{x}{The object to receive new class(es).}

\item{class}{\code{character(n)}. The name of additional class(es).}
}
\value{
An object with new classes.
}
\description{
Utility to \emph{prepend} a new class(es) to existing objects.
}
\examples{
class(iris)

add_class(iris, "new") |> class()

add_class(iris, c("A", "B")) |> class()    # 2 classes

add_class(iris, c("A", "data.frame")) |> class()    # no duplicates

add_class(iris, c("data.frame", "A")) |> class()    # re-orders if exists
}
\seealso{
\code{\link[=class]{class()}}, \code{\link[=typeof]{typeof()}}, \code{\link[=structure]{structure()}}
}
\author{
Stu Field
}
