\name{add_function_links_into_parsed}
\alias{add_function_links_into_parsed}
\title{Add Funciton Link...}
\usage{add_function_links_into_parsed(parser_output, source_link=FALSE)}

\description{
  Add Funciton Link Add the function link to the preparsed R code
}
\value{parsed output with functions with html links around them}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{parser_output}{pre-parsed output}
}
