\name{add_package_link_to_string}
\alias{add_package_link_to_string}
\title{Add package link to a string.}
\usage{add_package_link_to_string(string)}
\description{Add package link to a string.}
\keyword{internal}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\arguments{\item{string}{string to be evaluated}}
\examples{#add_package_link_to_string("quantreg, Hmisc, mapproj, maps, hexbin, gpclib, maptools")}
