% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_funs.R
\name{as.percentage}
\alias{as.perc}
\alias{as.percentage}
\title{Format a number as a percentage}
\usage{
as.percentage(num, digits = 1)

as.perc(num, digits = 1)
}
\description{
Format a number as a percentage
}
\section{Functions}{
\itemize{
\item \code{as.perc}: Shorthand
}}
\examples{

tmp <- runif(5)
as.percentage(tmp)

mtcars \%>\%
   count(gear) \%>\%
   mutate(prop=as.percentage(n/sum(n)))

}

