\name{author_email}
\alias{author_email}
\title{Author Email Parsing
parse the author email}
\usage{
  author_email(name, address)
}
\arguments{
  \item{name}{name of the email}

  \item{address}{email address}
}
\description{
  Author Email Parsing parse the author email
}
\author{
  Barret Schloerke
}
\keyword{internal}

