\name{body_text}
\alias{body_text}
\title{Extract source code of a function.}
\usage{
  body_text(package, fun)
}
\arguments{
  \item{fun}{name of function to get the source code from}
}
\value{
  NULL or source code of of the function
}
\description{
  Extract source code of a function.
}
\author{
  Barret Schloerke \email{schloerke@gmail.con} and Hadley
  Wickham
}
\keyword{internal}

