% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-ccc.R
\name{calc_ccc}
\alias{calc_ccc}
\title{Calculate Concordance Correlation Coefficient}
\usage{
calc_ccc(x, y)
}
\arguments{
\item{x}{A numeric vector of values to compare against \code{y}.}

\item{y}{A numeric vector of values to compare against \code{x}.}
}
\value{
A list of the following:
\item{rho.c:}{The concordance correlation coefficient (CCC).}
\item{ci95:}{The 95 percent confidence intervals of the CCC.}
\item{Z:}{The z-score of the CCC.}
\item{p.value:}{The p-value corresponding to the Z-score.}
}
\description{
Calculate the concordance correlation coefficient (CCC)
and it's significance value from two vectors of related data.
}
\examples{
calc_ccc(rnorm(100), rnorm(100))

v <- rnorm(100)
calc_ccc(v, v + 1)
}
\references{
Lawrence Lin, Biometrics (45): 255-268.
}
\seealso{
\code{\link[=cor]{cor()}}, \code{\link[=pnorm]{pnorm()}}
}
\author{
Stu Field
}
