% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-qvalue.R
\name{calc_qvalue}
\alias{calc_qvalue}
\alias{plot.q_value}
\title{Calculate q-values}
\usage{
calc_qvalue(
  p,
  lambda = seq(0, 0.95, 0.01),
  lambda.eval = NULL,
  match.storey = FALSE
)

\method{plot}{q_value}(x, ..., rng = c(0, 0.25))
}
\arguments{
\item{p}{A vector of p-values.}

\item{lambda}{A sequence of lambdas to evaluate.}

\item{lambda.eval}{A value of lambda to evaluate at, defaults to the maximum
value of the \code{lambda} sequence.}

\item{match.storey}{Logical. Should the output be matched to that of
Storey's qvalue() function?}

\item{x}{Object of class \code{q_value}.}

\item{...}{Additional arguments passed to the S3 generic method \code{plot}.}

\item{rng}{Numeric of \code{length == 2}. Range values.}
}
\value{
A list of class \code{q_value} containing:
\item{call}{The original call to \code{calc_qvalue()}.}
\item{p.value}{The original vector of p-values.}
\item{m}{?}
\item{lambda}{?}
\item{lambda.eval}{?}
\item{pi.lambda}{?}
\item{spline.fit}{?}
\item{pi0}{?}
\item{q.value}{A a vector of q-values.}

A plot
}
\description{
Calculates a vector of q-values corresponding to a vector of p-values.

S3 plot method for "q.value" class objects
}
\examples{
x <- withr::with_seed(101, c(seq(0.0001, 0.05, length = 50), runif(950)))
q <- calc_qvalue(x)

# S3 plot method
plot(q)
}
\references{
John Storey. PNAS. 2003.
}
\seealso{
\code{\link[=smooth.spline]{smooth.spline()}}
}
\author{
Stu Field
}
