% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-ss.R
\name{calc_ss}
\alias{calc_ss}
\title{Calculate the Sum of Squared Errors}
\usage{
calc_ss(x)
}
\arguments{
\item{x}{A numeric vector of data.}
}
\value{
The Sum of Squared Errors (SSE) of \code{x}.
}
\description{
Calculates the sum of the squared errors (SSE) for
a vector of numeric data.
}
\examples{
calc_ss(rnorm(100, 50, 5))
}
\author{
Stu Field
}
