\name{check_for_package}
\alias{check_for_package}
\title{check for a package...}
\usage{check_for_package(package)}
\description{check for a package
check to see if a package exists}
\keyword{internal}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\arguments{\item{package}{package in question}}
\examples{check_for_package("stats")
check_for_package("does_not_exist")}
