\name{code_info}
\alias{code_info}
\title{Find functions, counts, and links of given R text.}
\usage{
  code_info(parser_output)
}
\arguments{
  \item{parser_output}{text that has been parsed}
}
\value{
  data.frame containing the name, count and link of each
  function within the text
}
\description{
  Find functions, counts, and links of given R text.
}
\author{
  Barret Schloerke
}
\keyword{internal}

