% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colors.R
\name{color_chart}
\alias{color_chart}
\title{Show all colors of base R in a concise table.}
\usage{
color_chart(orderBy = c("number", "hue"))
}
\description{
Show all colors of base R in a concise table.
}
\seealso{
color_chart
}
