% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{color_scales}
\alias{color_scales}
\alias{scale_color_extreme}
\alias{scale_fill_extreme}
\title{Custom color scales for ggplot2}
\usage{
scale_color_extreme(...)

scale_fill_extreme(...)
}
\description{
A collection of color/fill scales for ggplot2.
}
\section{Functions}{
\itemize{
\item \code{scale_color_extreme}: A 'color' scale highlighting extreme values.

\item \code{scale_fill_extreme}: A 'fill' scale highlighting extreme values.
}}

