% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert2df.R
\name{convert2df}
\alias{convert2df}
\title{Convert R Objects to a Data Frame}
\usage{
convert2df(x)
}
\arguments{
\item{x}{Currently one of the following class objects:
\itemize{
\item \code{list}
\item \code{table}
\item \code{matrix}
\item \code{numeric}
\item \code{integer}
\item \code{factor}
\item \code{character}
}}
}
\value{
The original object \code{x} converted into a \code{data.frame} class.
}
\description{
Attempt to "smartly" convert an arbitrary object
into a \code{data.frame} object while keep the dimensions
of the same style.
}
\examples{
# Compare outputs:

# Character
char <- head(letters, 10)
char

as.data.frame(char)
convert2df(char)

# table
tab <- table(sample(c("A", "B"), 30, replace = TRUE))
tab

as.data.frame(tab)
convert2df(tab)

# matrix
mat <- matrix(1:9, ncol = 3L)
mat

as.data.frame(mat)
convert2df(mat)
}
\seealso{
\code{\link[=as.data.frame]{as.data.frame()}}
}
\author{
Stu Field
}
