% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-form.R
\name{create_form}
\alias{create_form}
\title{Create A Formula}
\usage{
create_form(response, features, collapse = c("+", "*"), env = NULL)
}
\arguments{
\item{response}{A quoted string representing the LHS of the
formula, i.e. the response variable (\code{Y}).}

\item{features}{A vector of quoted strings representing the
model features/predictors. Used to generate the right-hand
side (RHS) of the formula.}

\item{collapse}{\code{character(1)}. The separator for features.
Typically either \code{+} for main effects, \code{*} for interactions.}

\item{env}{The environment in which to evaluate the formula.
To keep the formula object as light-weight as possible,
the default is the current function environment/scope.
However, some occasions (e.g. unit testing) may require
the formula's environment to capture certain objects
necessary to fit a model (e.g. weights).}
}
\value{
A stats formula. See \code{\link[=formula]{formula()}}.
}
\description{
A convenient utility to quickly generate formula
objects in a standardized format and environment.
}
\examples{
ft <- c("ft_1234.56", "ft_8899.8", "ft_3334.3")
f1 <- create_form("class", ft)
print(f1, showEnv = FALSE)

# environment manipulation
# cleanest light-weight formula (default)
ls(environment(f1))

# capture another environment
e     <- new.env()
e$new <- LETTERS
e$obj <- rnorm(100)
f2    <- create_form("class", ft, env = e)
ls(environment(f2))
}
\seealso{
\code{\link[=eval]{eval()}}, \code{\link[=environment]{environment()}}
}
\author{
Stu Field
}
