% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-tab.R
\name{cross_tab}
\alias{cross_tab}
\title{Cross Tabulate Counts}
\usage{
cross_tab(data, ..., useNA = "ifany")
}
\arguments{
\item{data}{A \code{data.frame} (-like) object containing the
data variables from which counts are desired.}

\item{...}{A number (1 to 2) unquoted strings separated by commas.
Greater than 2 is generally not useful. Or additional
parameters passed to \code{\link[=table]{table()}}.}

\item{useNA}{See \code{\link[=table]{table()}}.}
}
\value{
A table of grouped counts based on splitting
variables with sums from each factor.
}
\description{
Create a contingency table of counts generated
by cross-classifying factors from groups
splitting on the values passed in the \code{...} argument.
A thin wrapper around \code{\link[=table]{table()}} but with simplifying
features. The sums of each row and column are added to the result.
}
\examples{
# 1 factor
cross_tab(mtcars, cyl)

# 2 factors
cross_tab(mtcars, cyl, gear)

# No quoted strings (!)
cross_tab(mtcars, "cyl")

# No external variable (!)
var <- "cyl"
cross_tab(mtcars, var)
}
\seealso{
\code{\link[=table]{table()}}, \code{\link[=addmargins]{addmargins()}}
}
\author{
Stu Field
}
