% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hmisc.R
\name{cut2}
\alias{cut2}
\title{Cut2 from Hmisc}
\usage{
cut2(x, cuts, m = 150, g, levels.mean = FALSE, digits, minmax = TRUE,
  oneval = TRUE, onlycuts = FALSE)
}
\description{
Shamelessly copied from Hmisc, which I don't like to load because of name collisions.
It not 'imported' from Hmisc because Hmisc's developpers are not bothered to deal with their namespace correctly.
}
\details{
#@seealso \link[Hmisc]{cut2}
}
