\name{dataframe_has_rows}
\alias{dataframe_has_rows}
\title{Determines whether or not the dataframe has rows.}
\usage{
  dataframe_has_rows(x)
}
\arguments{
  \item{x}{item in question}
}
\description{
  Determines whether or not the dataframe has rows.
}
\author{
  Barret Schloerke
}
\keyword{internal}

