% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dater.R
\name{dater}
\alias{dater}
\title{Standardize Date Format}
\usage{
dater(x = "\%Y-\%m-\%d")
}
\arguments{
\item{x}{An alternative format to the standard format.}
}
\value{
The current date in \code{YYYY-MM-DD} format (default).
}
\description{
A convenient wrapper to print the current date in a
consistent format for all package suite functionality.
}
\examples{
# with default format
dater()

# pass alternative format
dater("\%Y-\%m-\%d || \%H:\%M:\%S")
}
\seealso{
\code{\link[=format]{format()}}
}
