% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots-list2.R
\name{dots_list2}
\alias{dots_list2}
\title{Dynamic List Param Collection Rename Elements of a List}
\usage{
dots_list2(...)
}
\arguments{
\item{...}{Arguments to collect in a list.
These arguments are dynamic}
}
\value{
A named list corresponding to the
expressions passed to the \code{...} inputs.
}
\description{
Helper to collect the \code{...} params internally
passed that support non-standard evaluation
and splicing. See \code{rlang::dots_list()}
and \code{rlang::list2()}.
In general this is a stripped down version
to avoid the \code{rlang} import keeping the
package footprint minimal.
}
\examples{
foo <- function(...) {
  print(dots_list2(...))
  invisible()
}
foo(a = 2, b = 4)

foo(a = 2, b = 4, fun = function(x) mean(x))

foo(data.frame(a = 1, b = 2), data.frame(a = 8, b = 4))

# supports !!! splicing
args <- list(a = 2, b = 4, fun = function(x) mean(x))
foo(!!!args)
}
