% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich-test.R
\name{enrich_test}
\alias{enrich_test}
\alias{enrich_test.list}
\alias{enrich_test.data.frame}
\alias{enrich_test.matrix}
\title{Test for Enrichment}
\usage{
enrich_test(x, ...)

\method{enrich_test}{list}(x, ...)

\method{enrich_test}{data.frame}(x, ...)

\method{enrich_test}{matrix}(x, ...)
}
\arguments{
\item{x}{A \verb{2x2} confusion matrix (aka contingency table) containing
the binary decisions for each contingency (or a similarly structured
data frame). Can also be a (named) list containing each of the
4 contingencies. See examples.}

\item{...}{Arguments passed to \code{\link[stats:fisher.test]{stats::fisher.test()}}. In particular,
\code{alternative} which determines whether to check for both
enrichment and depletion ("two.sided") or specifically one
\emph{or} the other: enrichment = "greater"; depletion = "less".}
}
\value{
An "enrichment" class object, a list of the significance
tests calculated from a Hypergeometric Distribution
\code{\link[stats:Hypergeometric]{stats::dhyper()}} as well as those calculated via Fisher's Exact
\code{\link[stats:fisher.test]{stats::fisher.test()}} test for count data
testing the H_o that the odds ratio is equal to 1.
The p-values for various flavors of Hypergeometric test are:
\item{1-sided}{add here.}
\item{2-sided}{double of \code{1-sided}.}
\item{1-sided mid}{add description}
\item{2-sided mid}{double of \verb{1-sided mid}.}
\item{2-sided min lik}{this differs previous, but is most
similar to Fisher's Exact}
\item{2-sided min lik mid}{this is typically preferred}
}
\description{
Calculated whether \verb{2x2} table is enriched for a particular
group using Hypergeometric Distribution and
the Fisher's Exact test for count data.
}
\details{
Can also pass a \emph{named} list containing:
\tabular{rcl}{
\code{n11} \tab : \tab The corresponding \verb{[1,1]} position of the \verb{2x2} matrix.\cr
\code{n1_} \tab : \tab The sum of the top row of the table.\cr
\code{n_1} \tab : \tab The sum of the first column of the table.\cr
\code{n}   \tab : \tab The sum of the table.\cr
}
}
\examples{
c_mat <- matrix(c(4, 2, 3, 11), ncol = 2L)
enrich_test(c_mat)

# or pass a named list
en_list <- list(n11 = 4, n1_ = 7, n_1 = 6, n = 20)
enrich_test(en_list)
}
\seealso{
\code{\link[stats:Hypergeometric]{stats::dhyper()}}, \code{\link[stats:fisher.test]{stats::fisher.test()}}
}
\author{
Stu Field
}
