\name{eval_on_global}
\alias{eval_on_global}
\title{Eval text using the global environment.
Using memoise to speedup reproducability.}
\usage{
  eval_on_global(...)
}
\arguments{
  \item{txt}{text to be evaluated}
}
\description{
  Eval text using the global environment. Using memoise to
  speedup reproducability.
}
\author{
  Barret Schloerke
}
\keyword{evaluate}
\keyword{importFrom}
\keyword{internal}

