\name{eval_tag_output}
\alias{eval_tag_output}
\title{evaluate output tag...}
\usage{eval_tag_output(x)}
\description{evaluate output tag
tag the text to make sure it is considered output}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{x}{text to be tagged as output}}
