\name{evaluate_text}
\alias{evaluate_text}
\title{Evaluate text and return the corresponding text output and source.}
\usage{evaluate_text(txt, pic_base_name)}
\description{Evaluate text and return the corresponding text output and source.}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{txt}{text to be evaluated}
\item{pic_base_name}{base name for the picture files}}
