\name{first_item_pos}
\alias{first_item_pos}
\title{Find the first item position.}
\usage{first_item_pos(arr)}
\description{Find the first item position.}
\value{position of the first item to match "\\item" else 1}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{arr}{arr of items to look at}}
