\name{function_help_path}
\title{Return the help path of a function.}
\usage{function_help_path(func, source_link=FALSE)}
\description{Return the help path of a function.}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\alias{function_help_path}
\alias{function_help_path_mem}
\arguments{\item{x}{item to find the help path}
\item{source_link}{boolean to determine whether or not it is linking to the source or topic page}}
