\name{function_levels}
\alias{function_levels}
\title{Function levels.
go through the function text to find the function level (depth) of each function}
\usage{
  function_levels(text)
}
\arguments{
  \item{text}{text to be evaluated}
}
\description{
  Function levels. go through the function text to find the
  function level (depth) of each function
}
\author{
  Barret Schloerke
}
\keyword{internal}

