\name{functions_used}
\alias{functions_used}
\title{Functions used in the string of functions.}
\usage{
  functions_used(txt)
}
\arguments{
  \item{txt}{text in question.  should be full of
  functions}
}
\description{
  Functions used in the string of functions.
}
\author{
  Barret Schloerke
}
\keyword{internal}

