\name{get_datasets}
\alias{get_datasets}
\title{All Dataset Names...}
\usage{get_datasets(package)}
\description{All Dataset Names}
\value{\code{\link{vector}} of names that can be a dataset}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{package}{package to explore}}
