% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-outliers.R
\name{get_outliers}
\alias{get_outliers}
\title{Get Indices of Statistical Outliers}
\usage{
get_outliers(
  x,
  n_sigma = 3,
  mad_crit = 6,
  fold_crit = 5,
  type = c("nonparametric", "parametric")
)
}
\arguments{
\item{x}{`numeric(n). A vector of values to evaluate.}

\item{n_sigma}{\code{numeric(1)}. The the number of standard
deviations from the mean a \code{n.sigma} threshold for outliers.
Ignored if \code{type = "nonparametric"}.}

\item{mad_crit}{The median absolute deviation ("MAD")
criterion to use. Ignored if \code{type = "parametric"}.
Defaults to \code{(6 * mad)}.}

\item{fold_crit}{The fold-change criterion to evaluate.
Ignored if \code{type = "parametric"}. Defaults to \verb{5x}.}

\item{type}{\code{character(1)}. Matched. Either
"parametric" or "nonparametric" to determine
the type of outliers detection implementation.}
}
\value{
If "nonparametric": an integer vector of
indices corresponding to detected outliers.

If "parametric": an integer vector of indices with these additional attributes:
\item{mu}{the robustly fit mean (\eqn{\mu}).}
\item{sigma}{the robustly fit standard deviation (\eqn{\sigma}).}
\item{crit}{the 2 critical values beyond which a value is considered
an outlier.}
}
\description{
Calculates the indices of a vector of values that exceed a specified
statistical outlier criterion. This criterion is defined by differently
depending on the type of outlier detection implemented (see Details).
}
\section{outlier detection}{

There are 2 possible methods used to define an outlier measurement
and the return value depends on which method is implemented:
\enumerate{
\item The non-parametric case (default): agnostic to the distribution.
Outlier measurements are defined as falling outside \code{mad.crit * mad}
from the median \emph{and} a specified number of fold-changes from
the median (i.e. \code{fold.crit}; e.g. 5x).\cr
\strong{Note:} \code{n.sigma} is ignored.
\item The parametric case: the mean and standard deviation are
calculated robustly via \code{\link[=fit_gauss]{fit_gauss()}}. Outliers
are defined as measurements falling \emph{outside} +/- \code{n.sigma}*\eqn{\sigma}
from the the estimated \eqn{\mu}.\cr
\strong{Note:} \code{mad.crit} and \code{fold.crit} are ignored.
}
}

\examples{
withr::with_seed(101, {
  x <- rnorm(26, 15, 2)         # Gaussian
  x <- c(2, 2.5, x, 25, 25.9)   # add 4 outliers (2hi, 2lo)
})
get_outliers(x)                # non-parametric (default)
get_outliers(x, type = "para") # parametric
}
\seealso{
\code{\link[=fit_gauss]{fit_gauss()}}
}
\author{
Stu Field
}
