\name{get_servr_query_result}
\alias{get_servr_query_result}
\title{Solr query.
Retrieve a solr query}
\usage{
  get_servr_query_result(query_list, function_field =
  FALSE)
}
\arguments{
  \item{query_list}{list that contains the start position,
  query, and other parameters that are ready to be pasted}
}
\description{
  Solr query. Retrieve a solr query
}
\author{
  Barret Schloerke
}
\keyword{internal}

