\name{get_solr_query_result}
\alias{get_solr_query_result}
\title{Solr query.}
\usage{get_solr_query_result(query_list, function_field=FALSE)}
\description{Solr query.
Retrieve a solr query}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{query_list}{list that contains the start position, query, and other parameters that are ready to be pasted}}
