% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{group_at}
\alias{group_at}
\title{Group_by with regex based column selection}
\usage{
group_at(df, pattern)
}
\description{
Similar to `group_by` except grouping columns are selected using a regex.
}
