% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpr-package.R
\docType{package}
\name{helpr-package}
\alias{helpr}
\alias{helpr-package}
\title{helpr: Helper Utilities for R}
\description{
The helpr package contains numerous helpers, wrappers, and utilities used throughout the my analysis suite. It intentionally favors base R over higher level *tidyverse* to minimize imports.
}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

}
\keyword{internal}
\keyword{package}
