\name{helpr}
\alias{helpr}
\title{Helpr Documentation...}
\usage{helpr(installed=TRUE)}
\description{Helpr Documentation
Execute to show documentation}
\author{Hadley Wickham and Barret Schloerke \email{schloerke@gmail.com}}
\arguments{\item{installed}{use TRUE if the package is installed on from CRAN}}
