\name{helpr_package}
\title{Helpr Package...}
\usage{helpr_package(package)}

\description{
  Helpr Package
}
\alias{helpr_package}
\alias{helpr_package_mem}
\value{all the information necessary to produce a package site ("/package/:package/")}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
