\name{helpr_replay}
\alias{helpr_replay}
\title{Replay a list of evaluated results, just like you'd run them in a R...}
\usage{helpr_replay(x, pic_base_name)}
\description{Replay a list of evaluated results, just like you'd run them in a R
terminal.}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{x}{result from \code{\link{evaluate}}}
\item{pic_base_name}{base picture name to be used}}
