\name{helpr_replay_message}
\alias{helpr_replay_message}
\title{cat a 'replay'...}
\usage{helpr_replay_message(x)}
\description{cat a 'replay'
parse r code and print it}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{x}{result from \code{\link{evaluate}}}}
