\name{helpr_solr_search}
\alias{helpr_solr_search}
\title{Helpr Search.}
\usage{
  helpr_solr_search(query_list)
}
\arguments{
  \item{query_list}{list that contains the start position,
  query, and other parameters}
}
\value{
  returns all the necessary information from a search
}
\description{
  Helpr Search.
}
\author{
  Barret Schloerke
}
\keyword{internal}

