\name{index_all}
\alias{index_all}
\title{Index all packages.}
\usage{index_all(start_letter="a", verbose=TRUE)}
\description{Index all packages.
Index all packages into solr}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\arguments{\item{start_letter}{used when you want to start froma certain letter, such as 'q'}
\item{verbose}{should output be shown?}}
