\name{index_package}
\alias{index_package}
\title{Index package...}
\usage{index_package(package, start_letter="a", verbose=TRUE)}
\description{Index package
Index a whole package into solr}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\arguments{\item{package}{package in question}
\item{start_letter}{used when you want to start froma certain letter, such as 'q'}
\item{verbose}{should output be shown?}}
