% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{is_empty}
\alias{is_empty}
\title{Check file is empty}
\usage{
is_empty(file)
}
\arguments{
\item{file}{Character vector of filepath(s) to check.}
}
\value{
Logical vector
}
\description{
Check whether the size of the specified file on disk is exactly 0.
}
