\name{is_function}
\alias{is_function}
\title{Determines if object is a function...}
\usage{is_function(package, items)}
\description{Determines if object is a function}
\value{boolean}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{items}{items supplied from pkg_topics}}
