\name{is_section}
\alias{is_section}
\title{Determines whether or not tag is a section}
\usage{
  is_section(tag)
}
\arguments{
  \item{tag}{check to see if the tag is in a list of other
  tags}
}
\description{
  Determines whether or not tag is a section
}
\author{
  Hadley Wickham
}
\keyword{internal}

