% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{isFileEmpty}
\alias{isFileEmpty}
\title{Check if file is empty}
\usage{
isFileEmpty(file)
}
\arguments{
\item{file}{Character vector, path(s) to the file(s) to check.}
}
\description{
A vectorised wrapper around file.info which checks whether the size of the file on disk is exactly 0.
}

