% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagged-tbl.R
\name{jagged_tbl}
\alias{jagged_tbl}
\title{Create a Jagged Tibble}
\usage{
jagged_tbl(x)
}
\arguments{
\item{x}{A \emph{named} list of vectors with unequal lengths, to be
recast as a \code{tibble}.}
}
\value{
A \code{tibble} object with dimensions
\code{c(max(lengths(x)), length(x))}. Extra entries are
replaced with \code{NA}.
}
\description{
Combine a list of uneven (jagged) vectors into
one 2-dim object, \code{tibble} or \verb{data frame},
and fill in lengths with \code{NA}.
}
\examples{
p <- 5   # cols
n <- 10  # rows
x <- replicate(p, sample(1:1000, sample(1:n, 1))) |>
  set_Names(paste0("v", seq(p)))

tbl <- jagged_tbl(x)
tbl
}
\author{
Stu Field
}
