% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{keep_first}
\alias{keep_first}
\alias{keep_first.data.frame}
\alias{keep_first.default}
\title{Keep first value}
\usage{
keep_first(x)

\method{keep_first}{data.frame}(df)

\method{keep_first}{default}(df)
}
\description{
More or less equivalent to `keep_one`
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Method for data.frame

\item \code{default}: Default method
}}
\seealso{
keep_one
}

