% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_na.R
\name{keep_na_rows}
\alias{keep_na_rows}
\title{Keep only rows containing at least one NA}
\usage{
keep_na_rows(df)
}
\arguments{
\item{df}{Data.frame to process}
}
\description{
Useful for quick NA checks. Equivalent to `keep_na(df, everything())`
}
