% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{keep_one}
\alias{keep_one}
\title{Keep one obs only}
\usage{
keep_one(col, quiet = T)
}
\arguments{
\item{col}{A vector (typically the column to summarise)}

\item{quiet}{see details}
}
\description{
This is useful for use with dplyr's `summarize`, when one wants to keep in the summarized data
a column which do not need to be summarised.
}
\details{
The underlying assumption is that all values for the column considered are the same and
 we can just keep one when summarising.
 If there are different values, the first value of the vector is chosen and alternative values are ignored.
 If `quiet=FALSE`, a message is given every time this happens.

 Note that if `quiet=T`, no unicity check is performed and the first value is picked anyway.
}
\examples{
school_results <- data.frame(pupil=c("John", "John", "Mary", "Mary", "Henry"),
                             exam=c(1, 2, 1, 2, 1),
                             mark=c(9, 8, 10, 8, 6),
                             gender=c("M", "M", "F", "F", "M"))

pupils_results <- school_results \%>\%
                         group_by(pupil) \%>\%
                         summarise(av_mark=mean(mark),
                             gender=keep_one(gender))

}

