\name{last_ten_functions}
\alias{last_ten_functions}
\title{List last ten unique functions called from the command line.}
\usage{
  last_ten_functions(fun_list = get_function_history())
}
\arguments{
  \item{fun_list}{function list}
}
\value{
  fuctions that are from a package and their help paths
}
\description{
  List last ten unique functions called from the command
  line.
}
\author{
  Barret Schloerke
}
\keyword{internal}

