% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shinystan_detached.R
\name{launch_shinystan_detached}
\alias{launch_shinystan_detached}
\title{Launch shinystan as a detached R session}
\usage{
launch_shinystan_detached(object, port = getOption("shiny.port", 8080),
  quiet = FALSE)
}
\arguments{
\item{object}{The object to pass to launch_shinystan, typically a shinystan or stanfit object.}

\item{port}{Port that shinystan will bind. Uses the "shiny.port" option if set and defaults to 8080 otherwise.}

\item{quiet}{If TRUE, do not print anything.}
}
\description{
Launch shinystan as a detached R session
}
\details{
The `object` is first compressed and written to disk in a temporary file (see `tempfile`). 
A separate R session is started (using `system`) where the object is read and `launch_shinystan` is called.
When shinystan is closed, the temporary file is deleted.
}
\note{
Requires `shinystan` although it is not an explicit dependency.
}
\seealso{
shinystan::launch_shinystan
}
