% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_loading.R
\name{lib}
\alias{lib}
\alias{req}
\title{Load several packages in one go}
\usage{
lib(...)

req(...)
}
\description{
Packages should be given as characters.
}
\section{Functions}{
\itemize{
\item \code{req}: Same with `require`
}}

\examples{
\dontrun{lib("ggplot2", "RColorBrewer")}
}
