% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-mode.R
\name{lib_mode}
\alias{lib_mode}
\alias{lib_tree}
\alias{is_lib_mode}
\title{Activate and Deactivate Lib Mode}
\usage{
lib_mode(path = getOption("helpr_path"))

lib_tree()

is_lib_mode()
}
\arguments{
\item{path}{A directory location of an R library.}
}
\description{
When activated, \code{\link[=lib_mode]{lib_mode()}} adds a new library location
(aka \code{lib.loc}) to the top of the library tree.
This allows you to run code in an "sandbox", without
interfering with the other packages you have installed previously.
See \code{vignette("lib-mode")} for more details with examples.
}
\section{Functions}{
\itemize{
\item \code{lib_tree()}: A thin wrapper around \code{.libPaths()} to quickly
and easily view the \emph{current} library tree of
directories the session knows about.

\item \code{is_lib_mode()}: Determines whether a session is \emph{currently} in "lib mode".

}}
\examples{
\dontrun{
  dir.create("new-lib")
  lib_mode("new-lib")  # ON; activate
  lib_mode("new-lib")  # OFF; deactivate
  lib_mode("new-lib")  # ON; re-activate
}
}
\seealso{
\code{\link[=.libPaths]{.libPaths()}}
}
