\name{list_to_double_list}
\alias{list_to_double_list}
\title{Make a list into a nested list.
this is to be done to easily use sapply and keep the name of the item}
\usage{
  list_to_double_list(obj)
}
\arguments{
  \item{obj}{list to perform on}
}
\description{
  Make a list into a nested list. this is to be done to
  easily use sapply and keep the name of the item
}
\author{
  Barret Schloerke
}
\keyword{internal}

