\name{list_to_xml}
\alias{list_to_xml}
\title{Turn a list into a solr doc.}
\usage{list_to_xml(id, obj)}
\description{Turn a list into a solr doc.
turn a list into a solr doc}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{id}{id tag to be used}
\item{obj}{list to perform on}}
