\name{load_html}
\alias{load_html}
\title{Load an html page from the console.}
\usage{
  load_html(..., character.only = FALSE)
}
\arguments{
  \item{...}{site to be loaded}

  \item{character.only}{flag to substitute or not}
}
\description{
  Load an html page from the console.
}
\author{
  Barret Schloerke
}

