% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_parallel.R
\name{load_parallel}
\alias{load_parallel}
\title{Loads a parallel back-end}
\usage{
load_parallel(nbClust = 2, verbose = T)
}
\arguments{
\item{nbClust}{Number of nodes to create in the cluster.}
}
\description{
Builds a local cluster of nbClust nodes and registers it as parallel backend.
Will also try to stop any cluster previously created by the function.
}
