% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logicals.R
\name{logicals}
\alias{logicals}
\alias{is_int_vec}
\alias{is_logspace}
\alias{is_monotonic}
\alias{has_length}
\alias{len_one}
\alias{is_chr}
\alias{is_dbl}
\alias{is_lgl}
\alias{is_int}
\title{Logical Testing}
\usage{
is_int_vec(x)

is_logspace(x)

is_monotonic(x)

has_length(x)

len_one(x)

is_chr(x)

is_dbl(x)

is_lgl(x)

is_int(x)
}
\arguments{
\item{x}{A vector, data frame or \code{tibble} to be tested.}
}
\value{
Logical/boolean
}
\description{
These functional checks return boolean \code{TRUE/FALSE}
depending on the result of their test.
}
\section{Functions}{
\itemize{
\item \code{is_int_vec()}: A general test of whether a numeric
vector object contains only integer values.
This is a fix for the generally
undesired behavior of \code{\link[=is.integer]{is.integer()}} which
doesn't actually test for integer numbers (see its ?help).

\item \code{is_logspace()}: Checks if an object containing numeric data
is already in log space. This check assumes
proteomic values (RFU) and that the vector median,
or the entire data matrix, will be greater than
15 if in linear space and less than 10 if log10-transformed.

\item \code{is_monotonic()}: A general test of whether the numeric vector \code{x} is
monotonically \emph{increasing} or \emph{decreasing} in value.

\item \code{has_length()}: check that \code{length > 0}.

\item \code{len_one()}: check that length = 1.

\item \code{is_chr()}: check for scalar + character type.

\item \code{is_dbl()}: check for scalar + double type.

\item \code{is_lgl()}: check for scalar + logical type.

\item \code{is_int()}: check for scalar + integer type.

}}
\examples{
is_int_vec(10)         # does not return TRUE
is_int_vec(10L)        # does not return TRUE
is_int_vec(10)
is_int_vec(1:10)
is_int_vec(c(3.2, 1:10))
is_int_vec(rnorm(10))

# log-space
x <- rnorm(30, mean = 1000)
is_logspace(x)
is_logspace(log(x))

df <- data.frame(a = 1:5, ft_1234 = round(rnorm(5, mean = 5000, sd = 100), 1))
is_logspace(df)
df <- data.frame(a = 1:5, ft_3456 = round(rnorm(5, mean = 3, sd = 1), 1))
is_logspace(df)

# monotonic
is_monotonic(1:10)      # TRUE
is_monotonic(10:1)      # TRUE
is_monotonic(rnorm(10)) # FALSE
is_monotonic(seq(10, -10, by = -1)) # TRUE
}
\seealso{
\code{\link[=floor]{floor()}}, \code{\link[=is.integer]{is.integer()}}, \code{\link[=all]{all()}}, \code{\link[=is.numeric]{is.numeric()}}

\code{\link[=diff]{diff()}}, \code{\link[=all]{all()}}
}
\author{
Stu Field
}
