\name{make_add_xml}
\alias{make_add_xml}
\title{make it so the xml is an 'add'...}
\usage{make_add_xml(obj)}

\description{
  make it so the xml is an 'add' make it so the xml is an 'add' to be
  commited to solr
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{obj}{list to perform on}
}
