\name{make_field}
\alias{make_field}
\title{make a xml field...}
\usage{make_field(name, value)}

\description{
  make a xml field make a field for a solr document
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{name}{name of the field}
  \item{value}{value of the field}
}
