% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_CRI.R
\name{mean_CRI}
\alias{mean_CRI}
\title{A helpR function}
\usage{
mean_CRI(x, na.rm = FALSE, digits = 2, cred = 0.95)
}
\arguments{
\item{x}{A vector of numeric data to summarize}

\item{na.rm}{Remove NAs from x}

\item{digits}{Number of decimals to round to}

\item{cred}{Level of credible interval between 0 and 1.}
}
\description{
This function allows you to easily paste a mean and credible interval into a 
readable form for tables, graphs, text, etc...
}
\examples{
x <- rnorm(100)
mean_CRI(x)
}
