% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{A helpR function}
\usage{
mean_sd(x, na.rm = FALSE, digits = 2)
}
\arguments{
\item{x}{A vector of numeric data to summarize}

\item{na.rm}{Remove NAs from x}

\item{digits}{Number of decimals to round to}
}
\description{
This function allows you to easily paste a mean and sd into a 
readable form for tables, graphs, text, etc...
}
\examples{
x <- rnorm(100)
mean_sd(x)
}
