% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_IQR.R
\name{median_IQR}
\alias{median_IQR}
\title{A helpR function}
\usage{
median_IQR(x, na.rm = FALSE, digits = 2, IQR.type = 7)
}
\arguments{
\item{x}{A vector of numeric data to summarize}

\item{na.rm}{Remove NAs from x}

\item{digits}{Number of decimals to round to}

\item{IQR.type}{The type parameter to be passed to stats::IQR()}
}
\description{
This function allows you to easily paste a median and IQR into a 
readable form for tables, graphs, text, etc...
}
\examples{
x <- rnorm(100)
median_IQR(x)
}
