% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_management.R
\name{move}
\alias{move}
\title{Move objects to and from .rda files}
\usage{
move(listObjects, to, from = ".GlobalEnv", fn.object = FALSE,
  replace.objects = FALSE, move.doc.objects = TRUE, copy.only = FALSE,
  feedback = TRUE, map = TRUE)
}
\description{
Move objects to and from .rda files
}
\seealso{
rm.sv
}

