% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{msd}
\alias{msd}
\title{Mean plus-or-minus Standard Deviation}
\usage{
msd(x, na.rm = F, digits = 1, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{Whether to ignore NAs}

\item{digits}{Digits to use for rounding}

\item{...}{Passed to mean}
}
\description{
Returns a string in format "mean ± sd".
}
\seealso{
mqmmqm, fivenum
}

