% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{normalise}
\alias{normalise}
\title{Normalise a matrix by rows or column}
\usage{
normalise(mat, by = "col", norm = 1)
}
\arguments{
\item{mat}{Matrix to normalise}

\item{by}{"col" or "row"}

\item{norm}{1 (for L1 norm) or 2 (for L2 norm)}
}
\description{
Normalises rows or columns of a matrix according to the L1 or L2 norm.
}

