% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-value-FDR.R
\name{p_value_FDR}
\alias{p_value_FDR}
\alias{plot.fdr}
\title{Calculate Benjamini-Hochberg FDR Corrected Values}
\usage{
p_value_FDR(p, alpha = 0.05)

\method{plot}{fdr}(x, ...)
}
\arguments{
\item{p}{A numeric vector of p-values.}

\item{alpha}{\code{numeric(1)}. Significance level in \verb{[0, 1]}.}

\item{x}{A \code{fdr} object.}

\item{...}{Unused. For compliance with the \code{\link[=plot]{plot()}} generic default params.}
}
\value{
A \code{fdr} object. A \code{tibble} for the step-up procedure.
The \code{tibble} contains:
\item{p.value}{A sorted vector of the original p-values.}
\item{penalty (\code{k/m})}{The threshold value, corresponding to \code{k / m} (slope).}
\item{threshold}{The threshold value, corresponding to \code{alpha * k / m}.}
\item{p_hat}{Term in the minimum function corresponding to \code{p * m / k}.}
\item{fdr}{The FDR-adjusted p-values.}
\item{idx}{The indices of the original p-values.}
\item{fdr_p}{The FDR-adjusted p-values ordered by \code{idx}.}
\item{alpha}{The chosen significance threshold.}
}
\description{
Calculation and diagnostic plots to explain how B-H FDR
procedure works via Benjamini-Hochberg "step-up"
correction. FDR corrected p-values are calculated
from scratch and reported with other elements/steps
of the procedure.
}
\note{
You are basically solving for the slope (\code{k/m})
that makes the p-value (\code{alpha}) significant.
}
\examples{
p1  <- c(0.01, 0.013, 0.014, 0.19, 0.35, 0.5, 0.63, 0.67, 0.75, 0.81)
new <- p_value_FDR(p1)
new

old <- p.adjust(p1, method = "fdr")
all.equal(new$fdr_p, old)

plot(new)

p2 <- withr::with_seed(1001, runif(10, 0.0005, 0.3))
p_value_FDR(p2)

p3 <- withr::with_seed(669, c(runif(10), rep(0.5, 10)))
p_value_FDR(p3)

p4 <- withr::with_seed(101, c(runif(200), runif(25, 0.01, 0.1)))
p_value_FDR(p4)
}
\references{
Benjamini, Y., and Hochberg, Y. (1995). Controlling the false
discovery rate: a practical and powerful approach to multiple
testing. \emph{Journal of the Royal Statistical Society Series B}
\strong{57}, 289-300.

http://www.unc.edu/courses/2007spring/biol/145/001/docs/lectures/Nov12.html \cr
http://en.wikipedia.org/wiki/False_discovery_rate#Benjamini.E2.80.93Hochberg_procedure
}
\seealso{
\code{\link[=p.adjust]{p.adjust()}}, \code{\link[=cummin]{cummin()}}
}
\author{
Stu Field
}
