\name{package_and_topic_from_url}
\alias{package_and_topic_from_url}
\title{Pkg and topic from url...}
\usage{package_and_topic_from_url(url_txt)}

\description{
  Pkg and topic from url Retrieve the pkg and topic from the url
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{solr_param_string}{parameters that are to be passed onto solr}
}
