\name{package_description}
\alias{package_description}
\title{Package description...}
\usage{package_description(pkg, topic)}

\description{
  Package description
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{pkg}{package in question}
  \item{topic}{topic in question}
}
