\name{package_usage_functions}
\alias{package_usage_functions}
\title{Find all top level package usage functions...}
\usage{package_usage_functions(usage)}
\description{Find all top level package usage functions}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{usage}{usage in question}}
