\name{parse_help}
\alias{parse_help}
\title{Parse help file.
Function to turn a help topic into a convenient format.}
\usage{
  parse_help(rd, package)
}
\arguments{
  \item{rd}{item to be tagged recursively}
}
\value{
  item reformatted to be used in HTML
}
\description{
  Parse help file. Function to turn a help topic into a
  convenient format.
}
\author{
  Hadley Wickham and Barret Schloerke
}
\keyword{internal}

