\name{parse_item_list}
\alias{parse_item_list}
\title{Parse a group of "\\item" into a table with a bold item and reconstructed...}
\usage{parse_item_list(rd, package)}

\description{
  Parse a group of "\\item" into a table with a bold item and
  reconstructed description.
}
\value{table text with no attributes}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{rd}{item to be parsed}
  \item{package}{package looking at}
}
