\name{parse_pkg_desc_item}
\alias{parse_pkg_desc_item}
\title{Ensure package version is properly displayed (if not already in a nice
format).}
\usage{
  parse_pkg_desc_item(obj)
}
\description{
  Ensure package version is properly displayed (if not
  already in a nice format).
}
\author{
  Barret Schloerke
}
\keyword{internal}

