\name{parse_tabular}
\alias{parse_tabular}
\title{Parse a Tabular Section
Parse a tabular section to include the text alignments}
\usage{
  parse_tabular(tabular)
}
\arguments{
  \item{tabular}{rd item to parsed}
}
\description{
  Parse a Tabular Section Parse a tabular section to
  include the text alignments
}
\author{
  Barret Schloerke
}
\keyword{internal}

