\name{parse_text}
\alias{parse_text}
\title{Parse Text
wrapper to parser's \code{parser(text = text)}}
\usage{
  parse_text(text)
}
\arguments{
  \item{text}{text to be parsed}
}
\description{
  Parse Text wrapper to parser's \code{parser(text = text)}
}
\author{
  Barret Schloerke
}
\keyword{internal}

