\name{parse_usage}
\alias{parse_usage}
\title{Parse usage.
Parse the topic usage to add links to functions}
\usage{
  parse_usage(usage, package)
}
\arguments{
  \item{usage}{rd usage}
}
\description{
  Parse usage. Parse the topic usage to add links to
  functions
}
\author{
  Barret Schloerke
}
\keyword{internal}

