% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch.R
\name{patch}
\alias{patch}
\alias{patch.data.frame}
\alias{where}
\title{Conditional patching of data.frames}
\usage{
patch(object, cond, ...)

\method{patch}{data.frame}(object, cond, ..., quiet = FALSE)

where(..., .follow_symbols = FALSE)
}
\arguments{
\item{object}{Data.frame to patch}

\item{cond}{The condition(s) selecting the rows to patch. Should be wrapped in "where".}

\item{...}{Action(s) to take, in the form `var=new_value`.}

\item{quiet}{If FALSE, a message indicating number of patched rows will be printed.}
}
\description{
Performs one or several 'mutate' on a data.frame, but only on rows matching the specified condition(s).
}
\examples{
patch(mtcars, where(vs == 0, am == 1), gear = Inf, carb = carb + 10)

# Setting incorrect values in data$pos_var to NA
# Setting corresponding related_var observations to NA
data \%>\% patch(where(pos_var<0), pos_var=NA, related_var=NA)
}
\author{
Antoine Sachet
}

