\name{pkg_author_and_maintainers}
\alias{pkg_author_and_maintainers}
\title{Parse authors and maintainer.}
\usage{
  pkg_author_and_maintainers(authors, maintainer = NULL)
}
\arguments{
  \item{description}{list containing the \code{author} and
  \code{maintainer}}
}
\value{
  string containing the authors with links properly
  displayed
}
\description{
  Parse authors and maintainer.
}
\author{
  Barret Schloerke
}
\keyword{internal}

