\name{pkg_demos}
\alias{pkg_demos}
\title{Package Demos...}
\usage{pkg_demos(package)}

\description{
  Package Demos
}
\value{demo() information for the \code{demo_name} in \code{package}}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{package}{package name}
  \item{demo_name}{demo name}
}
