\name{pkg_internal_function_files}
\alias{pkg_internal_function_files}
\title{package internal functions...}
\usage{pkg_internal_function_files(package)}
\description{package internal functions
retrieve all package internal functions}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{package}{package in question}}
