\name{pkg_rddb_path}
\alias{pkg_rddb_path}
\title{Documentation Database Path...}
\usage{pkg_rddb_path(package)}

\description{
  Documentation Database Path
}
\value{\code{file.path} to the documentation database}
\keyword{internal}
\author{Hadley Wickham}
\arguments{
  \item{package}{package to explore}
}
