\name{pkg_topic}
\alias{pkg_topic}
\title{Package Topic R Documentation...}
\usage{pkg_topic(package, topic, file)}
\description{Package Topic R Documentation}
\author{Haldey Wickham}
\keyword{internal}
\value{text of the .rd file}
\arguments{\item{package}{package to explore}
\item{topic}{topic of the package to retrieve}
\item{file}{location of the rd database.  If it is \code{NULL}, it will be found.}}
