\name{pkg_topic_index_type}
\alias{pkg_topic_index_type}
\title{Determines object type...}
\usage{pkg_topic_index_type(package)}
\description{Determines object type}
\details{Requires the package to be loaded to accurately determine the type of the object}
\value{returns a type ("package", "dataset", "function", "internal", "help_name" or "NA")}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{package}{package to explore}}
