\name{pkg_topics_alias}
\alias{pkg_topics_alias}
\title{Topic Title and Aliases by Package...}
\usage{pkg_topics_alias(pkg)}

\description{
  Topic Title and Aliases by Package return information on the package,
  datasets, internal, and datasets
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{pkg}{package in question}
}
