\name{pkg_topics_index}
\alias{pkg_topics_index}
\title{Package topics alias to file index.}
\description{Package topics alias to file index.}
\value{\code{\link{data.frame}} containing \code{alias} (function name) and \code{file} that it is associated with}
\keyword{internal}
\author{Hadley Wickham}
\arguments{\item{package}{package to explore}}
