\name{pkg_topics_index_by_type}
\alias{pkg_topics_index_by_type}
\title{All Dataset Names...}
\usage{pkg_topics_index_by_type(package)}
\description{All Dataset Names}
\value{\code{\link{list}} of the items in \code{\link{pkg_topics_index}} as \code{function}, \code{dataset}, \code{internal}, \code{package}, or \code{NA}.}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{package}{package to explore}}
