\name{pkg_topics_rd}
\alias{pkg_topics_rd}
\title{Package Topics File Documentation...}
\description{Package Topics File Documentation}
\details{Items can be accessed by \emph{\code{list()}}\code{$file_name}}
\value{\code{\link{list}} containing the documentation file of each file of a package}
\keyword{internal}
\author{Hadley Wickham}
\arguments{\item{package}{package to explore}}
