\name{pkg_vigs}
\alias{pkg_vigs}
\title{List all package vignettes.}
\usage{pkg_vigs(package)}
\description{List all package vignettes.}
\value{\code{subset} of the \code{vignette()$results} \code{data.frame} ("Package", "LibPath", "Item" and "Title")}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{package}{package to explore}}
