% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_hist}
\alias{plot_hist}
\title{Convenience function to plot a histogram with a pipe-compatible interface.}
\usage{
plot_hist(df, var, main = paste("Histogram of", var), xlab = var, ...)
}
\arguments{
\item{df}{A data.frame contaning the data}

\item{var}{A string naming the variable to plot}

\item{main}{Title of the plot}

\item{xlab}{Label of the x-axis}

\item{...}{Arguments passed to hist()}
}
\description{
Convenience function to plot a histogram with a pipe-compatible interface.
}
\examples{
 mtcars \%>\% plot_hist("mpg")
}
