% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_funs.R
\name{pretty_colnames}
\alias{pretty_colnames}
\title{Format the column names of a data.frame for pretty printing}
\usage{
pretty_colnames(df)
}
\description{
Format the column names of a data.frame for pretty printing
}

