\name{print.help_files_with_topic}
\alias{print.help_files_with_topic}
\title{Route to the topics with multiple files to correct page.}
\usage{
  \method{print}{help_files_with_topic} (x, ...)
}
\arguments{
  \item{x}{path to help}

  \item{...}{other arguments ignored}
}
\description{
  Modified from \code{utils:::print.help_files_with_topic}
}
\author{
  Barret Schloerke
}

