% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_bar.R
\name{progress_bar}
\alias{progress_bar}
\alias{reset_progress_bar}
\alias{update_progress_bar}
\title{Progress bar facility}
\usage{
reset_progress_bar(loop_data, title = NULL, by = "", ...)

update_progress_bar(value = NULL, ...)
}
\description{
Straightforward progress bar based on `textProgressBar`, suitable for use with a linear foreach.
}
\section{Functions}{
\itemize{
\item \code{reset_progress_bar}: Prepare the progress bar to loop over some data

\item \code{update_progress_bar}: Update the progress bar value
}}

\examples{
\dontrun{
reset_progress_bar(big_vec)
foreach(vec=big_vec) \%do\% {
  # ...
  update_progress_bar()
  }
}
}
