\name{read_url}
\alias{read_url}
\title{Read URL.
Retrieve the text from a URL}
\usage{
  read_url(url_string)
}
\arguments{
  \item{url_string}{url to explore}
}
\value{
  plain text from that url
}
\description{
  Read URL. Retrieve the text from a URL
}
\author{
  Barret Schloerke
}
\keyword{internal}

