% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{rearrange}
\alias{rearrange}
\title{Arrange rows by variable}
\usage{
rearrange(df, ...)
}
\description{
Wrapper for arranging a data.frame with respect to some columns
and moving said columns to the front of the data.frame.
}
\examples{

# TRUE
}
\seealso{
dplyr::arrange
}

