% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{remove_na}
\alias{remove_na}
\title{Remove NA's from selected columns}
\usage{
remove_na(df, ...)
}
\description{
This is a selective version of `na.omit` which lets you choose which columns to consider.
It is a wrapper around dplyr's `select` and `complete.cases`.
}
\seealso{
keep_na
}

