% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{rename_col}
\alias{rename_col}
\title{Rename a column programmatically}
\usage{
rename_col(x, old, new)
}
\arguments{
\item{x}{data.frame}

\item{old}{Character: column to rename. Can be a regex.}

\item{new}{Character: new name for the column}
}
\description{
Shortcomings in `dplyr::rename_` forces to have this handy function.
Simple wrapper around `names<-` and `gsub`.
}
\examples{
df <- data.frame(a=1:5, b=6:10)
col_to_change <- "a"
df \%>\% rename_col(col_to_change, "c")

}

