% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_funs.R
\name{rename_levels}
\alias{rename_levels}
\title{Easily rename one or more of the existing levels of fac}
\usage{
rename_levels(fac, old, new)
}
\arguments{
\item{fac}{Factor for which levels should be changed}

\item{old}{Character vector of old levels to rename}

\item{new}{Character vector of new names to replace those in `old`.}
}
\value{
`fac` with modified levels.
}
\description{
Easily rename one or more of the existing levels of fac
}
\details{
`new` should be as long as `old` or of length 1.
If it is as long as `old`, then the first value in `new` will replace the first value in `old`, etc.
if it is of length 1, then all levels in `old` will be replace by `new`.

This works as expected even if a level in `old` is not present in fac or if a level in `new` is already present in fac.
}
\examples{
# creating a factor
fac <- factor(c("PC", "laptop", "mobile", "telephone", "TV"))

# 1 replacement
rename_levels(fac, old="PC", new="computer")

# 1 replacement to already existing level
rename_levels(fac, old="telephone", new="mobile")

# 2 parallel replacements
rename_levels(fac, old=c("PC", "TV"), new=c("computer", "television"))

# 2 replacement to one common level
rename_levels(fac, old=c("PC", "laptop"), new=c("computer"))

# Replacement of inexisting level
rename_levels(fac, old="minitel", new="oldie")
}
