\name{render_json}
\alias{render_json}
\title{Render JSON...}
\usage{render_json(object)}
\description{Render JSON}
\details{Produce JSON from an R object}
\author{Hadley Wickham and Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{object}{object to be turned into JSON}}
