% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep-lgl.R
\name{rep_lgl}
\alias{rep_lgl}
\title{All Values Same}
\usage{
rep_lgl(x)
}
\arguments{
\item{x}{A vector of values. Can be one of:
\code{numeric}, \code{character}, \code{factor}, or \code{logical}.
If the vector is type \code{double} and NA values are present,
they are first removed.}
}
\value{
\code{logical(1)}.
}
\description{
Returns a boolean testing if all values of a vector
are identical. This function does NOT compare two independent
vectors. Use \code{isTRUE(all.equal())} for such comparisons.
}
\examples{
rep_lgl(1:4)

rep_lgl(rep(5, 10))

rep_lgl(rep("A", 10))

rep_lgl(letters)

rep_lgl(c(TRUE, TRUE, TRUE))

rep_lgl(c(TRUE, TRUE, FALSE))
}
\seealso{
\code{\link[=rep]{rep()}}, \code{\link[=isTRUE]{isTRUE()}}, \code{\link[=all.equal]{all.equal()}}
}
\author{
Stu Field
}
