\name{retrieve_usage_functions_and_methods}
\alias{retrieve_usage_functions_and_methods}
\title{Find all usage functions and methods...}
\usage{retrieve_usage_functions_and_methods(usage)}
\description{Find all usage functions and methods}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{usage}{usage in question}}
