% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_management.R
\name{rm.sv}
\alias{rm.sv}
\title{Remove object from any location on the searchpath}
\usage{
rm.sv(list, env.on.path = "", pos = NA, rm.doc.objects = FALSE,
  feedback = TRUE)
}
\description{
Remove object from any location on the searchpath
}

