\name{rp}
\title{Router Info...}
\description{Router Info}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\alias{rp}
\alias{set_router_info}
\alias{set_fouter_file_path}
\alias{set_router_custom_route}
\alias{router_info}
\alias{fouter_file_path}
\alias{router_custom_route}
\keyword{internal}
