\name{safely_order_funcs}
\alias{safely_order_funcs}
\title{Order functions safely by name.}
\usage{safely_order_funcs(vect)}
\description{Order functions safely by name.}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{vect}{string vector to be processed}}
