\name{save_picture}
\alias{save_picture}
\title{Save a picture...}
\usage{save_picture(obj_name, obj_plot)}
\description{Save a picture
save a picture into the temp directory}
\value{the path to the picture (using the website)}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
