% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-rds.R
\name{save_rds}
\alias{save_rds}
\alias{save_rda}
\alias{getCompression}
\title{Save Compressed \verb{*.rds} File}
\usage{
save_rds(object, file)

save_rda(..., file)

getCompression(file)
}
\arguments{
\item{object}{\R object to serialize.}

\item{file}{a \link[base]{connection} or the name of the file where the \R object
    is saved to or read from.}

\item{...}{the names of the objects to be saved (as symbols or
    character strings).}
}
\value{
Returns \code{file}, invisibly.
}
\description{
A thin wrapper around \code{\link[=saveRDS]{saveRDS()}} that ensures two things:
\itemize{
\item The path extension is consistent and lowercase \verb{*.rds}
\item The compression used is \code{"xz"}, which is often optimal
for our data types
}
}
\section{Functions}{
\itemize{
\item \code{save_rda()}: similar to \code{\link[=save_rds]{save_rds()}}, but for saving serialized \verb{*.rda} compressed files.

\item \code{getCompression()}: determine the type of compression for a serialized binary file.

}}
\examples{
\dontrun{
# all are the same
save_rds(x, "outfile.rds")
save_rds(x, "outfile.RDS")
save_rds(x, "outfile.Rds")

# similar functionality for rda
# as with `save()`, must specify file argument explicitly
save_rda(x, file = "outfile.rda")

# determine the compression ('xz')
getCompression("outfile.rda")
}
}
