\name{save_xml}
\alias{save_xml}
\title{Save page info.
Save page info into xml for solr}
\usage{
  save_xml(txt, file_name = tempfile())
}
\arguments{
  \item{txt}{xml text string}

  \item{file_name}{location to save the file. Defaults to a
  temp file that is discarded when R shuts down.}
}
\description{
  Save page info. Save page info into xml for solr
}
\author{
  Barret Schloerke
}
\keyword{internal}

