\name{search_query_path}
\alias{search_query_path}
\title{search query path...}
\usage{search_query_path(query, start_pos)}
\description{search query path
return a html path for a search}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{query}{query to be used}
\item{start_pos}{postion to start at}}
