\name{send_commit_command}
\alias{send_commit_command}
\title{Send commit command to Solr
Send a commit command to Solr to finalized any submissions}
\description{
  Send commit command to Solr Send a commit command to Solr
  to finalized any submissions
}
\author{
  Barret Schloerke
}
\keyword{internal}

