\name{servr_similar}
\alias{servr_similar}
\title{Similar pages.
Find related pages and return info in a data.frame}
\usage{
  servr_similar(topic)
}
\arguments{
  \item{topic}{title to be used to find similar results}
}
\description{
  Similar pages. Find related pages and return info in a
  data.frame
}
\author{
  Barret Schloerke
}
\keyword{internal}

