% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_compare.R
\name{set_compare}
\alias{set_compare}
\title{Compare elements from 2 sets}
\usage{
set_compare(vec1, vec2)
}
\arguments{
\item{vec1}{Vector 1}

\item{vec2}{Vector 2}
}
\description{
Create a 2x2 matrix showing the counts of common and different elements
}
\examples{
set_compare(1:5, 4:10)

}
