% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Names.R
\name{set_Names}
\alias{set_Names}
\title{Set Names of an Object}
\usage{
set_Names(x, nms = x, ...)
}
\arguments{
\item{x}{The object to name.}

\item{nms}{The names to apply to \code{x}. Recycled as necessary.}

\item{...}{Arguments passed to \code{nms} if a function.}
}
\description{
Similar to \code{\link[stats:setNames]{stats::setNames()}}, but with more features.
You can specify names in the following ways:
\itemize{
\item a character vector (recycled if \code{length(nms) != length(x)})
\item a function to be applied to names of \code{x}
\item via \code{...} if not passing a function, \code{c(nms, ...)}
}
}
\examples{
set_Names(head(letters), head(LETTERS)) # apply names

set_Names(head(letters))                # names self

set_Names(head(letters), toupper)       # apply fn

set_Names(set_Names(head(letters)), NULL)     # NULL always removes names

set_Names(head(letters), "foo")               # repeat to length

set_Names(head(letters), c("foo", "bar"))     # repeat to length

set_Names(head(letters), "foo", "bar", "baz") # repeat to length via `...`
}
