% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colors.R
\name{show_colors}
\alias{show_colors}
\title{Display the codes for colors in R}
\usage{
show_colors(start.num = 1, Labels = T, num.row = 15, num.col = 4)
}
\arguments{
\item{start.num}{the number of the first color to be displayed.  The  default setting displays the colors in 60 color increments.}

\item{Labels}{T if color labels are desired; F to omit.}

\item{num.row}{the number of rows in the color display.}

\item{num.col}{the number of columns in the color display.}
}
\description{
Adapted from http://research.stowers-institute.org/efg/R/Color/Chart/index.htm.
}
\seealso{
color_chart
}

