% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle.R
\name{shuffle}
\alias{shuffle}
\alias{shuffle.data.frame}
\alias{shuffle.default}
\title{Shuffle data}
\usage{
shuffle(x, ...)

\method{shuffle}{data.frame}(df)

\method{shuffle}{default}(x)
}
\description{
Reorders a data structure randomly.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Method for data.frame

\item \code{default}: Default method for vectors or list
}}

