\name{solr_similar}
\alias{solr_similar}
\title{Similar pages.}
\usage{solr_similar(topic)}
\description{Similar pages.
Find related pages and return info in a data.frame}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{topic}{title to be used to find similar results}}
