\name{solr_topic}
\alias{solr_topic}
\title{#' Produce a message that states Solr is not running.}
\usage{solr_topic(package, topic)}
\description{#' Produce a message that states Solr is not running.
#'
#'}
\author{Barret Schloerke \email{schloerke@gmail.com}
#'}
\keyword{internal}
\keyword{Make}
\keyword{Helpr}
\keyword{topic}
\keyword{into}
\keyword{xml}
\keyword{for}
\keyword{solr.}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{package}{package to use}
\item{topic}{topic to explore}}
