% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbl.R
\name{symbl}
\alias{symbl}
\alias{show_symbols}
\title{A list of UTF-8 or ASCII symbols.}
\format{
A named list, see \code{names(symbl)}.
}
\usage{
symbl

show_symbols()
}
\description{
Similar to the \pkg{cli} package listing of UTF-8/Unicode symbols
and their ASCII fall-backs depending on the environment they are called
into. Can be controlled by the option \code{options(cli.unicode = T/F)} if set,
otherwise the \code{\link[=l10n_info]{l10n_info()}}\verb{$UTF-8} local information is used to detect
UTF-8 support.
}
\details{
\code{show_symbols()} is a convenient print output to list
of various available UTF-8 (or ASCII) symbols to the screen.
}
\examples{
# available symbols
show_symbols()

# ascii versions
withr::with_options(list(cli.unicode = FALSE, width = 80L), show_symbols())
}
