\name{tag_link}
\alias{tag_link}
\title{link to another topic}
\usage{
  tag_link(fun, pkg = NULL, topic_page = fun)
}
\arguments{
  \item{fun}{function to tag}

  \item{pkg}{package to look in}

  \item{topic_page}{topic page of the function}
}
\description{
  link to another topic
}
\author{
  Hadley Wickham and Barret Schloerke
}
\keyword{internal}

