\name{top_ten_functions}
\alias{top_ten_functions}
\title{List top 10 most frequently called functions.}
\usage{
  top_ten_functions(fun_list = get_function_history())
}
\arguments{
  \item{fun_list}{function list}
}
\value{
  fuctions that are from a package and their help paths
}
\description{
  List top 10 most frequently called functions.
}
\author{
  Barret Schloerke
}
\keyword{internal}

