\name{topic_is_internal}
\alias{topic_is_internal}
\title{Internal topic function.}
\usage{topic_is_internal(help)}
\description{Internal topic function.}
\value{boolean}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{help}{\code{pkg_topic(}\emph{\code{topic}}\code{)}  is checked to see if a keyword is "internal"}}
