% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit-help.R
\name{expect_error_free}
\alias{expect_error_free}
\alias{skip_on_check}
\alias{skip_on_test}
\alias{skip_on_jenkins}
\title{Unit Test Helpers}
\usage{
expect_error_free(...)

skip_on_check()

skip_on_test()

skip_on_jenkins()
}
\arguments{
\item{...}{Arguments passed to \code{\link[testthat:expect_no_error]{testthat::expect_no_error()}}.}
}
\description{
Manage whether certain tests are skipped during
\code{\link[devtools:check]{devtools::check()}}, \code{\link[devtools:test]{devtools::test()}}, or during a temporary automated
pipeline builds. This may be desired if certain tests are time consuming
or special environment characteristics make them undesirable.
See \code{\link[testthat:skip]{testthat::skip_if()}} and \code{\link[testthat:skip]{testthat::skip_on_cran()}} for details.
}
