\name{update_packs}
\alias{update_packs}
\title{Update all packages that are old and currently loaded or installed.}
\usage{update_packs(all=FALSE)}

\description{
  Update all packages that are old and currently loaded or installed.
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
