\name{urlJSON_to_list}
\alias{urlJSON_to_list}
\title{URL with JSON result to list.}
\usage{
  urlJSON_to_list(url_string)
}
\arguments{
  \item{url_string}{url that contains a JSON output to be
  turned into a list}
}
\description{
  URL with JSON result to list.
}
\author{
  Barret Schloerke
}
\keyword{internal}

