% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical_funs.R
\name{vswitch}
\alias{vswitch}
\title{Vectorized switch operation}
\usage{
vswitch(EXPR, ...)
}
\arguments{
\item{EXPR}{Expression to switch on.}

\item{...}{Switching rules passed to switch}
}
\description{
The `switch` operator cannot be applied to a vector.
`vswitch` removes this limitation.
}
\examples{
x <- letters[1:5]
vswitch(x, a="first", b="second", ">3")
}

