% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write.table.gz}
\alias{write.table.gz}
\alias{write.csv.gz}
\title{Write directly to a compressed gz file}
\usage{
write.table.gz(x, file, ...)

write.csv.gz(x, file, ...)
}
\description{
Write directly to a compressed gz file
}
\section{Functions}{
\itemize{
\item \code{write.csv.gz}: @export
}}

