% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-list.R
\name{write_list}
\alias{write_list}
\title{Write List to CSV File}
\usage{
write_list(x, file, rn_title = NULL, append = FALSE, ...)
}
\arguments{
\item{x}{A list to be written to file, typically a list of data frames.}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing.  \code{""} indicates output to the console.}

\item{rn_title}{Title for the row names column (for data frames).
Must match the length of the \code{x} argument. Defaults to empty \code{""}.}

\item{append}{logical. Only relevant if \code{file} is a character
    string.  If \code{TRUE}, the output is appended to the
    file.  If \code{FALSE}, any existing file of the name is destroyed.}

\item{...}{Additional arguments passed to \code{\link[=write.table]{write.table()}}.}
}
\value{
The list object, invisibly.
}
\description{
Write a list of data frames of vectors (or a mixture
of both) to a single \verb{*.csv} file.
}
\examples{
tmp <- lapply(LETTERS[1:5], function(x) rnorm(10, mean = 10, sd = 3))
names(tmp) <- LETTERS[1:5]
write_list(tmp, file = tempfile(fileext = ".csv"))

# with a data frame in list
tmp$mtcars <- head(mtcars, 10)
write_list(tmp, file = tempfile(fileext = ".csv"))
}
\seealso{
\code{\link[=write.table]{write.table()}}, \code{\link[=file]{file()}}
}
\author{
Stu Field
}
