\name{add_function_links_into_parsed}
\alias{add_function_links_into_parsed}
\title{Add funciton link.
Add the function link to the preparsed R code}
\usage{
  add_function_links_into_parsed(parser_output, source_link
  = FALSE)
}
\arguments{
  \item{parser_output}{pre-parsed output}
}
\value{
  parsed output with functions with html links around them
}
\description{
  Add funciton link. Add the function link to the preparsed
  R code
}
\author{
  Barret Schloerke
}
\keyword{internal}

