\name{add_package_link_to_string}
\alias{add_package_link_to_string}
\title{Add package link to a string.}
\usage{
  add_package_link_to_string(string)
}
\arguments{
  \item{string}{string to be evaluated}
}
\description{
  Add package link to a string.
}
\examples{
#add_package_link_to_string("quantreg, Hmisc, mapproj, maps, hexbin, gpclib, maptools")
}
\author{
  Barret Schloerke
}
\keyword{internal}

