\name{author_email}
\alias{author_email}
\title{Author Email Parsing...}
\usage{author_email(name, address)}
\description{Author Email Parsing
parse the author email}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{name}{name of the email}
\item{address}{email address}}
