\name{body_text}
\alias{body_text}
\title{Extract source code of a function.}
\usage{body_text(package, fun)}

\description{
  Extract source code of a function.
}
\value{NULL or source code of of the function}
\author{Barret Schloerke \email{schloerke@gmail.con} and Hadley Wickham}
\keyword{internal}
\arguments{
  \item{fun}{name of function to get the source code from}
}
