\name{body_text}
\alias{body_text}
\title{Function Source...}
\usage{body_text(package, fun)}

\description{
  Function Source Work out the source code of a function.
}
\value{NULL or source code of of the function}
\author{Barret Schloerke \email{schloerke@gmail.con} and Hadley Wickham}
\keyword{internal}
\arguments{
  \item{fun}{function to get the source code from}
}
