% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-qvalue.R
\name{calc_qvalue}
\alias{calc_qvalue}
\alias{plot.q_value}
\title{Calculate q-values}
\usage{
calc_qvalue(
  p,
  lambda = seq(0, 0.95, 0.01),
  lambda_eval = NULL,
  match_storey = FALSE
)

\method{plot}{q_value}(x, ..., rng = c(0, 0.25))
}
\arguments{
\item{p}{A vector of p-values.}

\item{lambda}{A sequence of lambdas to evaluate.}

\item{lambda_eval}{A value of lambda to evaluate at,
defaults to the maximum value of the \code{lambda} sequence.}

\item{match_storey}{\code{logical(1)}. Should the output match
the \code{qvalue()} function exactly?}

\item{x}{A \code{q_value} class object.}

\item{...}{Additional arguments passed to the S3 plot generic.}

\item{rng}{\code{numeric(2)}. Range of values.}
}
\value{
A list of class \code{q_value} containing:
\item{call}{The original call to \code{calc_qvalue()}.}
\item{p_value}{The original vector of p-values.}
\item{m}{?}
\item{lambda}{?}
\item{lambda_eval}{?}
\item{pi_lambda}{?}
\item{spline_fit}{?}
\item{pi0}{?}
\item{q_value}{A a vector of q-values.}

A cool plot.
}
\description{
Calculates a vector of q-values corresponding
to a vector of p-values.
}
\examples{
x <- withr::with_seed(101, c(seq(0.0001, 0.05, length = 50), runif(950)))
q <- calc_qvalue(x)

# S3 plot method
plot(q)
}
\references{
John Storey. PNAS. 2003.
}
\seealso{
\code{\link[=smooth.spline]{smooth.spline()}}
}
\author{
Stu Field
}
