\name{code_info}
\alias{code_info}
\title{Find functions, counts, and links of given R text...}
\usage{code_info(parser_output)}

\description{
  Find functions, counts, and links of given R text
}
\value{data.frame containing the name, count and link of each function within the text}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{parser_output}{text that has been parsed}
}
