% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-tab.R
\name{cross_tab}
\alias{cross_tab}
\title{Cross Tabulate Counts}
\usage{
cross_tab(x, ..., useNA = "ifany")
}
\arguments{
\item{x}{A \code{data.frame} (-like) object containing the
data variables from which counts are desired.}

\item{...}{A number of unquoted expressions or quoted strings
separated by commas. Generally greater than 3 is not useful.}

\item{useNA}{See \code{\link[=table]{table()}}.}
}
\value{
A table of grouped counts based on splitting
variables with sums from each factor.
}
\description{
Create a contingency table of counts generated
by cross-classifying factors from groups
splitting on the values passed in the \code{...} argument.
The sums of each row and column are added to the result.
}
\examples{
# 1 factor
cross_tab(mtcars, cyl)

# quoted string
cross_tab(mtcars, "cyl")

# with external variable
var <- "cyl"
cross_tab(mtcars, var)

# 2 factors
cross_tab(mtcars, cyl, gear)

cross_tab(mtcars, "cyl", "gear")

cross_tab(mtcars, c("cyl", "gear"))

# 3 factors
cross_tab(mtcars, cyl, gear, am)
}
\seealso{
\code{\link[=table]{table()}}, \code{\link[=addmargins]{addmargins()}}
}
\author{
Stu Field
}
