\name{dataframe_has_rows}
\alias{dataframe_has_rows}
\title{Dataframe has Rows...}
\usage{dataframe_has_rows(x)}

\description{
  Dataframe has Rows determines whether or not the dataframe has rows
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{x}{item in question}
}
