\name{dataframe_has_rows}
\alias{dataframe_has_rows}
\title{Determines whether or not the dataframe has rows.}
\usage{dataframe_has_rows(x)}

\description{
  Determines whether or not the dataframe has rows.
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{x}{item in question}
}
