\name{demo_info}
\alias{demo_info}
\title{Demo information.}
\usage{
  demo_info(package, demo_name)
}
\arguments{
  \item{package}{package name}

  \item{demo_name}{demo name}
}
\value{
  demo() information for the \code{demo_name} in
  \code{package}
}
\description{
  Demo information.
}
\author{
  Barret Schloerke
}
\keyword{internal}

