\name{demo_info}
\alias{demo_info}
\title{Demo Information...}
\usage{demo_info(package, demo_name)}
\description{Demo Information}
\value{demo() information for the \code{demo_name} in \code{package}}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{package}{package name}
\item{demo_name}{demo name}}
