\name{demo_src_file}
\alias{demo_src_file}
\title{Demo Source File...}
\usage{demo_src_file(package, demo_name)}

\description{
  Demo Source File
}
\value{"" if the demo is not found for that package, file.path if the demo folder is found}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{package}{package name}
  \item{demo_name}{demo name}
}
