\name{demo_src_file}
\alias{demo_src_file}
\title{Demo source file.}
\usage{
  demo_src_file(package, demo_name)
}
\arguments{
  \item{package}{package name}

  \item{demo_name}{demo name}
}
\value{
  "" if the demo is not found for that package, file.path
  if the demo folder is found
}
\description{
  Demo source file.
}
\author{
  Barret Schloerke
}
\keyword{internal}

