% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff-vecs.R
\name{diff_vecs}
\alias{diff_vecs}
\title{Diff Two Vectors}
\usage{
diff_vecs(x, y, verbose = interactive())
}
\arguments{
\item{x}{First vector to compare.}

\item{y}{Second vector to compare.}

\item{verbose}{Should diff summary be printed directly to the console?
Is \code{TRUE} for interactive sessions.}
}
\value{
An invisible list is returned with the set diffs of each vector.
The elements of the list are:
\item{unique_x:}{Entries unique to the \code{x} vector.}
\item{unique_y:}{Entries unique to the \code{y} vector.}
\item{inter:}{The intersect of the two vectors.}
\item{unique:}{All unique entries of the union of the two vectors.}
}
\description{
A convenient diff tool to determine
how and where two character vectors differ.
}
\examples{
diff_vecs(LETTERS[1:10], LETTERS[5:15])

diffs <- diff_vecs(LETTERS[1:10], LETTERS[5:15], verbose = FALSE)
diffs
}
\seealso{
\code{\link[=setdiff]{setdiff()}}, \code{\link[=union]{union()}}, \code{\link[=intersect]{intersect()}}
}
\author{
Stu Field
}
