\name{exec_example}
\alias{exec_example}
\title{evaluate example...}
\usage{exec_example(package, topic)}
\description{evaluate example
stole from example.  It didn't have the option of character.only}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{package}{package in question}
\item{topic}{topic in question}}
