\name{exec_pkg_demo}
\alias{exec_pkg_demo}
\title{Evaluate demo in the R console.}
\usage{
  exec_pkg_demo(package, dem)
}
\arguments{
  \item{package}{package in question}

  \item{dem}{demo in question}
}
\description{
  Evaluate demo in the R console.
}
\author{
  Barret Schloerke
}
\keyword{internal}

