\name{exec_pkg_demo}
\alias{exec_pkg_demo}
\title{evaluate demo...}
\usage{exec_pkg_demo(package, dem)}
\description{evaluate demo
evaluate demo in the R console}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{package}{package in question}
\item{dem}{demo in question}}
