% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-find.R, R/filesystem.R
\name{file_find}
\alias{file_find}
\alias{filesystem}
\alias{helpr_path}
\alias{as_helpr_path}
\alias{is.dir}
\alias{ls_dir}
\alias{info_dir}
\alias{file_ext}
\alias{file_ext<-}
\alias{rm_file_ext}
\alias{set_file_ext}
\alias{as_helpr_bytes}
\title{Determine Path to File}
\usage{
file_find(pattern, root = Sys.getenv("HOME"))

helpr_path(..., ext = "")

as_helpr_path(x)

is.dir(x)

ls_dir(dir = ".", regexp = NULL, all = FALSE, ...)

info_dir(dir = ".", ...)

file_ext(file)

file_ext(file) <- value

rm_file_ext(file)

set_file_ext(file, ext)

as_helpr_bytes(x)
}
\arguments{
\item{pattern}{\code{character(1)}. The pattern as a regex.}

\item{root}{The path of the root directory to start
the highest level search.}

\item{...}{Character vectors to construct paths, \code{length == 1}
values are recycled as appropriate too complete pasting.
Alternatively, arguments passed to \code{\link[=dir]{dir()}} (for \code{\link[=ls_dir]{ls_dir()}}).}

\item{ext, value}{An optional extension to append to the generated path.}

\item{regexp}{A regular expression, e.g. "\verb{[.]csv$}",
see the \code{pattern} argument to \code{\link[=dir]{dir()}}. Files are
collated according to \code{"C"} locale rules, so that
they are ordered consistently with \code{\link[fs:dir_ls]{fs::dir_ls()}}.}

\item{all}{If \code{TRUE} hidden files are also returned.}

\item{file, x, dir}{A file-system location, directory, or path.
Vectorized paths are allowed where possible.}
}
\value{
The path(s) containing the regular expression
specified in \code{pattern}.
}
\description{
Determine the path of a defined file via brute force search of
(optionally) the root directory.

Generalized file system utilities and operations customized for
data structures.
It was heavily influenced by the
\pkg{fs} package: \url{https://github.com/r-lib/fs}.
}
\section{Functions}{
\itemize{
\item \code{helpr_path()}: Construct a path to a file.

\item \code{as_helpr_path()}: Coerce to a \code{helpr_path} object.

\item \code{is.dir()}: Test if location is a directory.

\item \code{ls_dir()}: List the directory contents.

\item \code{info_dir()}: Lists the directory contents similar to \code{ls -l}.

\item \code{file_ext()}: Extracts the file extension from a file path.

\item \code{file_ext(file) <- value}: Replaces an existing extension. See \code{\link[=set_file_ext]{set_file_ext()}}.

\item \code{rm_file_ext()}: Removes the file extension from a file path.

\item \code{set_file_ext()}: Replaces the existing file extension with \code{ext}.
Extensions of \code{length == 1} are recycled.

\item \code{as_helpr_bytes()}: Coerce to a \code{helpr_bytes} object.

}}
\examples{

# wrapper around `ls_dir()`
\dontrun{
  file_find("Makefile", ".")
  file.create("myfile.txt")
  file_find("myfile.txt", "..")
  unlink("myfile.txt", force = TRUE)
}
# paths
helpr_path("foo", "bar", "baz")    # no ext

helpr_path("foo", "bar", "baz", ext = "zip")  # ext

helpr_path("foo", letters[1:3], ext = "txt")  # recycled args

# directories
ls_dir()

is.dir(ls_dir())

info_dir()

# extensions
file_ext("foo.txt")

rm_file_ext("foo/bar.txt")

set_file_ext("foo/bar.csv", "tsv")
set_file_ext(c("foo.txt", NA, "bar.csv"), "R")   # NAs unchanged & 'R' recycled

x <- "foo.txt"
file_ext(x) <- "csv"
x
}
