\name{first_item_pos}
\alias{first_item_pos}
\title{Find the first item position.}
\usage{
  first_item_pos(arr)
}
\arguments{
  \item{arr}{arr of items to look at}
}
\value{
  position of the first item to match "\\item" else 1
}
\description{
  Find the first item position.
}
\author{
  Barret Schloerke
}
\keyword{internal}

