\name{function_and_link}
\alias{function_and_link}
\title{Return the package functions and links of a given text.}
\usage{function_and_link(parser_output)}

\description{
  Return the package functions and links of a given text.
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{parser_output}{text that has been parsed}
}
