\name{function_help_path}
\alias{function_help_path}
\alias{function_help_path_mem}
\title{Return the help path of a function.}
\usage{
  function_help_path(func, source_link = FALSE)
}
\arguments{
  \item{x}{item to find the help path}

  \item{source_link}{boolean to determine whether or not it
  is linking to the source or topic page}
}
\description{
  Return the help path of a function.
}
\author{
  Barret Schloerke
}
\keyword{internal}

