\name{function_levels}
\alias{function_levels}
\title{function levels...}
\usage{function_levels(text)}

\description{
  function levels go through the function text to find the function
  level (depth) of each function
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{text}{text to be evaluated}
}
