\name{get_solr_query_result}
\alias{get_solr_query_result}
\title{Solr Query...}
\usage{get_solr_query_result(solr_param_string)}

\description{
  Solr Query Retrieve a solr query
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{solr_param_string}{parameters that are to be passed onto solr}
}
