% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/give-praise.R
\name{give_praise}
\alias{give_praise}
\title{Praise User}
\usage{
give_praise()
}
\description{
Part of collaboration is being friendly to your users.
Praise your colleagues with encouragement when they
use your code appropriately and when it runs error free!
This is a wrapper around the \pkg{praise} package and
takes no arguments. Additional flexibility may be added
in the future. Output can be suppressed by setting the
\code{"signal.quiet"} option (see example).
}
\examples{
# random praise 1
give_praise()

# random praise 2
give_praise()

# suppress praise
withr::with_options(list(signal.quiet = TRUE), give_praise())
}
