\name{helpr}
\alias{helpr}
\title{Start helpr.}
\usage{
  helpr(launch_browser = TRUE, no_internetz = NULL, custom
  = NULL)
}
\arguments{
  \item{launch_browser}{should this function launch a
  browser for immediate viewing of help pages?}

  \item{no_internetz}{should this function allow the
  browser to source internet files and actions? Default is
  to allow internet connections}

  \item{custom}{should this function launch in the custom
  httpd handlers env? Default is to be at the root of ""}
}
\description{
  Start helpr.
}
\author{
  Barret Schloerke and Hadley Wickham
}

