\name{helpr}
\alias{helpr}
\title{Start helpr.}
\usage{helpr(launch_browser=TRUE, no_internetz=FALSE, custom=FALSE)}
\description{Start helpr.}
\author{Barret Schloerke \email{schloerke@gmail.com} and Hadley Wickham}
\arguments{\item{launch_browser}{should this function launch a browser for immediate viewing of help pages?}
\item{no_internetz}{should this function allow the browser to source internet files and actions?}
\item{custom}{should this function launch in the custom httpd handlers env?}}
