\name{helpr}
\alias{helpr}
\title{Start helpr.}
\usage{helpr(launch_browser=TRUE, .local_mode=FALSE)}
\description{Start helpr.}
\author{Hadley Wickham and Barret Schloerke \email{schloerke@gmail.com}}
\arguments{\item{launch_browser}{should this function launch a browser for immediate
viewing of help pages?}}
