\name{helpr_function}
\alias{helpr_function}
\title{Helpr Function...}
\usage{helpr_function(package, func)}

\description{
  Helpr Function Render all the information to display a topic source
  page
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{package}{package in question}
  \item{func}{function in question}
}
