\name{helpr_package}
\alias{helpr_package}
\alias{helpr_package_mem}
\title{Package information.}
\usage{
  helpr_package(package)
}
\value{
  all the information necessary to produce a package site
  ("/package/:package/")
}
\description{
  Package information.
}
\author{
  Barret Schloerke
}
\keyword{internal}

