\name{helpr_replay}
\title{Replay a list of evaluated results.}
\usage{helpr_replay(x, pic_base_name)}
\description{Replay a list of evaluated results.
Replay a list of evaluated results, just like you'd run them in a R
terminal.}
\alias{helpr_replay}
\alias{helpr_replay.list}
\alias{helpr_replay.character}
\alias{helpr_replay.source}
\alias{helpr_replay.warning}
\alias{helpr_replay.message}
\alias{helpr_replay.error}
\alias{helpr_replay.value}
\alias{helpr_replay.recordedplot}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{x}{result from \code{\link{evaluate}}}
\item{pic_base_name}{base picture name to be used}}
