\name{helpr_solr_search}
\alias{helpr_solr_search}
\title{Helpr Search.}
\usage{helpr_solr_search(query_list)}
\description{Helpr Search.}
\value{returns all the necessary information from a search}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{query_list}{list that contains the start position, query, and other parameters}}
