\name{highlight}
\alias{highlight}
\title{Highlight R Text...}
\usage{highlight(parser_output, source_link=FALSE)}

\description{
  Highlight R Text Highlights R text to include links to all functions
  and make it easier to read
}
\value{highlighted text}
\author{Hadley Wickham and Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{parser_output}{text to be parsed and highlighted}
}
