\name{highlight}
\alias{highlight}
\title{Highlight R text.
Highlights R text to include links to all functions and make it easier to read}
\usage{
  highlight(parser_output, source_link = FALSE)
}
\arguments{
  \item{parser_output}{text to be parsed and highlighted}
}
\value{
  highlighted text
}
\description{
  Highlight R text. Highlights R text to include links to
  all functions and make it easier to read
}
\author{
  Hadley Wickham and Barret Schloerke
}
\keyword{internal}

