\name{index_all}
\alias{index_all}
\alias{index_packages}
\title{Index all packages.
Index all packages into solr}
\usage{
  index_all(start_letter = "a", verbose = TRUE)
}
\arguments{
  \item{start_letter}{used when you want to start froma
  certain letter, such as 'q'}

  \item{verbose}{should output be shown?}
}
\description{
  Index all packages. Index all packages into solr
}
\author{
  Barret Schloerke
}

