% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_first.R
\name{keep_first}
\alias{keep_first}
\alias{keep_first.data.frame}
\alias{keep_first.list}
\alias{keep_first.default}
\title{Keep first value}
\usage{
keep_first(x)

\method{keep_first}{data.frame}(df)

\method{keep_first}{list}(df)

\method{keep_first}{default}(df)
}
\description{
Useful when piping. Can be used in `summarise`.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Method for data.frame (keeps first row)

\item \code{list}: Method for list

\item \code{default}: Default method
}}

