% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{keep_na}
\alias{keep_na}
\title{Keep only rows with NA's in selected columns}
\usage{
keep_na(df, ...)
}
\description{
This is the opposite of remove_na.
Only rows with at least one NA in at least one of the selected columns are kept in the data.frame.
}
\seealso{
remove_na
}
