\name{last_item_pos}
\alias{last_item_pos}
\title{Find the last Item position.}
\usage{last_item_pos(arr)}
\description{Find the last Item position.}
\value{position of the last item to match "\\item" else 0}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{arr}{arr of items to look at}}
