\name{last_item_pos}
\alias{last_item_pos}
\title{Find the last Item position.}
\usage{
  last_item_pos(arr)
}
\arguments{
  \item{arr}{arr of items to look at}
}
\value{
  position of the last item to match "\\item" else 0
}
\description{
  Find the last Item position.
}
\author{
  Barret Schloerke
}
\keyword{internal}

