\name{last_ten_functions}
\alias{last_ten_functions}
\title{List last ten unique functions called from the command line.}
\usage{last_ten_functions(fun_list=get_function_history())}

\description{
  List last ten unique functions called from the command line.
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\value{fuctions that are from a package and their help paths}
\arguments{
  \item{fun_list}{function list}
}
