\name{last_ten_functions}
\alias{last_ten_functions}
\title{Last Ten Functions...}
\usage{last_ten_functions(fun_list=get_function_history())}
\description{Last Ten Functions
return the last ten unique functions that were executed}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\value{fuctions that are from a package and their help paths}
\arguments{\item{fun_list}{function list}}
