\name{list_to_double_list}
\alias{list_to_double_list}
\title{make a list into a nested list...}
\usage{list_to_double_list(obj)}

\description{
  make a list into a nested list this is to be done to easily use
  sapply and keep the name of the item
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{obj}{list to perform on}
}
