\name{list_to_xml}
\alias{list_to_xml}
\title{Turn a list into a solr doc.
turn a list into a solr doc}
\usage{
  list_to_xml(id, obj)
}
\arguments{
  \item{id}{id tag to be used}

  \item{obj}{list to perform on}
}
\description{
  Turn a list into a solr doc. turn a list into a solr doc
}
\author{
  Barret Schloerke
}
\keyword{internal}

