\name{make_add_xml}
\alias{make_add_xml}
\title{Make it so the xml is an 'add'.
make it so the xml is an 'add' to be commited to solr}
\usage{
  make_add_xml(obj)
}
\arguments{
  \item{obj}{list to perform on}
}
\description{
  Make it so the xml is an 'add'. make it so the xml is an
  'add' to be commited to solr
}
\author{
  Barret Schloerke
}
\keyword{internal}

