\name{make_field}
\alias{make_field}
\title{Make a xml field.
make a field for a solr document}
\usage{
  make_field(name, value)
}
\arguments{
  \item{name}{name of the field}

  \item{value}{value of the field}
}
\description{
  Make a xml field. make a field for a solr document
}
\author{
  Barret Schloerke
}
\keyword{internal}

