\name{make_field}
\alias{make_field}
\title{Make a xml field.}
\usage{make_field(name, value)}
\description{Make a xml field.
make a field for a solr document}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{name}{name of the field}
\item{value}{value of the field}}
