% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture-em.R
\name{normal_k2_mixture}
\alias{normal_k2_mixture}
\alias{plot.mix_k2}
\title{2-distribution (k2) Gaussian Mixture Model}
\usage{
normal_k2_mixture(
  data,
  pars = list(start_mu = NULL, start_sd = NULL, start_pi = NULL),
  max_iter = 1000,
  max_restarts = 25,
  eps = 1e-08
)

\method{plot}{mix_k2}(x, type = c("density", "likelihood", "posterior"), title = NULL, ...)
}
\arguments{
\item{data}{\code{numeric(n)}.}

\item{pars}{Initial values for \code{start_mu}, \code{start_sd}, and \code{start_pi}.}

\item{max_iter}{\code{integer(1)}. Max number of iterations to perform.}

\item{max_restarts}{\code{integer(1)}. Max number of restarts ro perform.}

\item{eps}{\code{double(1)}. Machine precision for when to stop the algorithm.}

\item{x}{A \code{mix_k2} object generated from \code{normal_k2_mixture}.}

\item{type}{\code{character(1)}. Matched string one of: "density",
"likelihood" or "posterior".}

\item{title}{\code{character(1)}. Title for the plot.}

\item{...}{Additional parameters for extensibility.}
}
\value{
A \code{mix_k2} class object.
}
\description{
Estimates the parameters of a 2 distribution mixture model
via expectation maximization.

S3 plot method for \code{"mix_k2"} objects.
}
\examples{
# Generate 2 gaussian distributions
x <- withr::with_seed(101,
  c(rnorm(100, mean = 10, sd = 1), rnorm(100, mean = 2, sd = 2)))
mix_theta <- normal_k2_mixture(x)
mix_theta

plot(mix_theta)
plot(mix_theta, "like")
plot(mix_theta, "post")
}
\references{
Tibshirani and Hastie

See Tibshirani and Hastie ("bible"); pg. 273.
}
\author{
Stu Field
}
