% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\alias{normalize}
\title{Normalise a matrix by rows or column}
\usage{
normalise(mat, norm = 1, by = "col")

normalize(mat, norm = 1, by = "col")
}
\arguments{
\item{mat}{Matrix to normalise}

\item{norm}{1 (for L1 norm) or 2 (for L2 norm)}

\item{by}{"col" or "row"}
}
\description{
Normalise rows or columns of a matrix for the L1 or L2 norm.
}
\section{Functions}{
\itemize{
\item \code{normalize}: Alternative spelling
}}

