\name{package_and_topic_from_url}
\alias{package_and_topic_from_url}
\title{Pkg and topic from URL.}
\usage{package_and_topic_from_url(url_txt)}
\description{Pkg and topic from URL.
Retrieve the pkg and topic from the URL}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{url_txt}{url to be parsed}}
