\name{parse_help}
\alias{parse_help}
\title{Parse Help File...}
\usage{parse_help(rd, package)}
\description{Parse Help File
Function to turn a help topic into a convenient format.}
\value{item reformatted to be used in HTML}
\author{Hadley Wickham and Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{rd}{item to be tagged recursively}}
