\name{parse_item_list}
\alias{parse_item_list}
\title{Parse a group of "\\item" into a table with a bold item and reconstructed
description.}
\usage{
  parse_item_list(rd, package)
}
\arguments{
  \item{rd}{item to be parsed}

  \item{package}{package looking at}
}
\value{
  table text with no attributes
}
\description{
  Parse a group of "\\item" into a table with a bold item
  and reconstructed description.
}
\author{
  Barret Schloerke
}
\keyword{internal}

