\name{parse_item_list}
\alias{parse_item_list}
\title{Parse Item List...}
\usage{parse_item_list(rd)}
\description{Parse Item List
parse a group of "\\item" into a table with a bold item and reconstructed description}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{rd}{item to be parsed
return table text with no attributes}}
