\name{parse_pkg_desc_item}
\alias{parse_pkg_desc_item}
\title{Parse a package description...}
\usage{parse_pkg_desc_item(obj)}

\description{
  Parse a package description makes sure that a package version is
  properly displayed if it is not given in a nice format
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
