\name{parse_tabular}
\alias{parse_tabular}
\title{Parse a Tabular Section...}
\usage{parse_tabular(tabular)}
\description{Parse a Tabular Section
Parse a tabular section to include the text alignments}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{tabular}{rd item to parsed}}
