\name{pkg_and_topic_from_help_url}
\alias{pkg_and_topic_from_help_url}
\title{Retrieve the package and the topic from a url that contains both.}
\usage{pkg_and_topic_from_help_url(url_string)}
\description{Retrieve the package and the topic from a url that contains both.}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{url_string}{url in question}}
