\name{pkg_author_and_maintainers}
\alias{pkg_author_and_maintainers}
\title{Parse Authors and Maintainer...}
\usage{pkg_author_and_maintainers(authors, maintainer)}
\description{Parse Authors and Maintainer}
\value{string containing the authors with links properly displayed}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{description}{list containing the \code{author} and \code{maintainer}}}
