\name{pkg_demos}
\alias{pkg_demos}
\title{List all demos in a package.}
\usage{
  pkg_demos(package)
}
\arguments{
  \item{package}{package name}
}
\value{
  demo() information for the \code{demo_name} in
  \code{package}
}
\description{
  List all demos in a package.
}
\author{
  Barret Schloerke
}
\keyword{internal}

