\name{pkg_demos}
\alias{pkg_demos}
\title{List all demos in a package.}
\usage{pkg_demos(package)}

\description{
  List all demos in a package.
}
\value{demo() information for the \code{demo_name} in \code{package}}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{package}{package name}
}
