\name{pkg_rddb_path}
\alias{pkg_rddb_path}
\title{Documentation database path.}
\usage{
  pkg_rddb_path(package)
}
\arguments{
  \item{package}{package to explore}
}
\value{
  \code{file.path} to the documentation database
}
\description{
  Documentation database path.
}
\author{
  Hadley Wickham
}
\keyword{internal}

