\name{pkg_topic}
\alias{pkg_topic}
\title{Package topic R documentation.}
\usage{
  pkg_topic(package, topic, file = NULL)
}
\arguments{
  \item{package}{package to explore}

  \item{topic}{topic of the package to retrieve}

  \item{file}{location of the rd database.  If it is
  \code{NULL}, it will be found.}
}
\value{
  text of the .rd file
}
\description{
  Package topic R documentation.
}
\author{
  Haldey Wickham and Barret Schloerke
}
\keyword{internal}

