\name{pkg_topics_alias}
\alias{pkg_topics_alias}
\title{Topic title and aliases by package.
return information on the package, datasets, internal, and datasets}
\usage{
  pkg_topics_alias(pkg)
}
\arguments{
  \item{pkg}{package in question}
}
\description{
  Topic title and aliases by package. return information on
  the package, datasets, internal, and datasets
}
\author{
  Barret Schloerke
}
\keyword{internal}

