\name{pkg_topics_alias}
\alias{pkg_topics_alias}
\title{Topic title and aliases by package.}
\usage{pkg_topics_alias(pkg)}
\description{Topic title and aliases by package.
return information on the package, datasets, internal, and datasets}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{pkg}{package in question}}
