\name{pkg_topics_index}
\alias{pkg_topics_index}
\title{Package Topics Alias to File Index...}
\description{Package Topics Alias to File Index}
\value{\code{\link{data.frame}} containing \code{alias} (function name) and \code{file} that it is associated with}
\keyword{internal}
\author{Hadley Wickham}
\arguments{\item{package}{package to explore}}
