\name{pkg_topics_index}
\alias{pkg_topics_index}
\title{Package topics alias to file index.}
\usage{
  pkg_topics_index(...)
}
\arguments{
  \item{package}{package to explore}
}
\value{
  \code{\link{data.frame}} containing \code{alias}
  (function name) and \code{file} that it is associated
  with
}
\description{
  Package topics alias to file index.
}
\author{
  Hadley Wickham
}
\keyword{internal}

