\name{pkg_topics_rd}
\alias{pkg_topics_rd}
\title{Package topics file documentation.}
\usage{
  pkg_topics_rd(...)
}
\arguments{
  \item{package}{package to explore}
}
\value{
  \code{\link{list}} containing the documentation file of
  each file of a package
}
\description{
  Items can be accessed by
  \emph{\code{list()}}\code{$file_name}
}
\author{
  Hadley Wickham
}
\keyword{internal}

