\name{print.help_files_with_topic}
\alias{print.help_files_with_topic}
\title{help files with topic for helpr...}
\usage{print.help_files_with_topic(x, ...)}

\description{
  help files with topic for helpr route to the correct website
}

\details{
  function taken from utils
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{x}{path to help}
  \item{...}{other arguments ignored}
}
