\name{print.help_files_with_topic}
\alias{print.help_files_with_topic}
\title{Route to the topics with multiple files to correct page.}
\usage{print.help_files_with_topic(x, ...)}

\description{
  Route to the topics with multiple files to correct page.
}

\details{
  Modified from \code{utils:::print.help_files_with_topic}
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{x}{path to help}
  \item{...}{other arguments ignored}
}
