\name{reconstruct}
\alias{reconstruct}
\title{Recursively reconstruct R documentation.}
\usage{
  reconstruct(rd, package = NULL)
}
\arguments{
  \item{rd}{rd in question}
}
\description{
  Recursively reconstruct R documentation.
}
\author{
  Hadley Wickham and Barret Schloerke
}
\keyword{internal}

