\name{reconstruct}
\alias{reconstruct}
\title{Recursively reconstruct R documentation.}
\usage{reconstruct(rd, package)}

\description{
  Recursively reconstruct R documentation.
}
\author{Hadley Wickham and Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{rd}{rd in question}
}
