\name{reconstruct}
\alias{reconstruct}
\title{Reconstruct R Documentation...}
\usage{reconstruct(rd, package)}

\description{
  Reconstruct R Documentation recursively reconstruct R documentation
}
\author{Hadley Wickham and Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{rd}{rd in question}
}
