\name{render_snippet}
\alias{render_snippet}
\title{Render HTML Snippet.}
\usage{render_snippet(template, params=list(), path=helpr_path())}
\description{Render HTML Snippet.}
\value{text rendered from the template}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\arguments{\item{template}{name of file to use in the folder "views"}
\item{params}{list containing objects to be sent to the template}
\item{path}{path to be used to find the "views" folder}}
