\name{safely_order_funcs}
\alias{safely_order_funcs}
\title{Order Functions Safely by Name...}
\usage{safely_order_funcs(vect)}
\description{Order Functions Safely by Name}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{vect}{string vector to be processed}}
