\name{save_xml}
\alias{save_xml}
\title{Save page info...}
\usage{save_xml(txt, file_name=tempfile())}

\description{
  Save page info Save page info into xml for solr
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{txt}{xml text string}
  \item{file_name}{location to save the file. Defaults to a temp file that is discarded when R shuts down.}
}
