\name{search_query_path}
\alias{search_query_path}
\title{Search query path.
return a html path for a search}
\usage{
  search_query_path(query = "example", start_pos = 0)
}
\arguments{
  \item{query}{query to be used}

  \item{start_pos}{postion to start at}
}
\description{
  Search query path. return a html path for a search
}
\author{
  Barret Schloerke
}
\keyword{internal}

