\name{send_system_command}
\alias{send_system_command}
\title{Send system command to Solr...}
\usage{send_system_command(system_string)}

\description{
  Send system command to Solr Send a system command to Solr to add /
  update files to Solr
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
