\name{solr_delete_package}
\alias{solr_delete_package}
\title{Deletes everything that belongs to a package in solr.}
\usage{solr_delete_package(package)}
\description{Deletes everything that belongs to a package in solr.}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\arguments{\item{package}{package to be deleted}}
