\name{solr_delete_package}
\alias{solr_delete_package}
\title{Delete a Package on Solr...}
\usage{solr_delete_package(package)}

\description{
  Delete a Package on Solr Deletes everything that belongs to a package
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\arguments{
  \item{package}{package to be deleted}
}
