\name{solr_similar}
\alias{solr_similar}
\title{Similar Pages...}
\usage{solr_similar(topic_title)}

\description{
  Similar Pages Find related pages and return info in a data.frame
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{topic_title}{title to be used to find similar results}
}
