\name{solr_topic}
\alias{solr_topic}
\title{helpr topic into xml for solr...}
\usage{solr_topic(package, topic)}

\description{
  helpr topic into xml for solr
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{package}{package to use}
  \item{topic}{topic to explore}
}
