\name{solr_topic}
\alias{solr_topic}
\title{Make Helpr topic into xml for solr.}
\usage{
  solr_topic(package, topic)
}
\arguments{
  \item{package}{package to use}

  \item{topic}{topic to explore}
}
\description{
  Make Helpr topic into xml for solr.
}
\author{
  Barret Schloerke
}
\keyword{internal}

