\name{tag_simple}
\alias{tag_simple}
\title{tag a simple item
tag an item that is contained in the simple tags list}
\usage{
  tag_simple(tag, text)
}
\arguments{
  \item{tag}{tag looking for}

  \item{text}{text to go in the tag}
}
\description{
  tag a simple item tag an item that is contained in the
  simple tags list
}
\author{
  Hadley Wickham
}
\keyword{internal}

