\name{ten_functions}
\alias{ten_functions}
\title{Both the top ten most called and last ten functions.}

\description{
  Both the top ten most called and last ten functions.
}
\value{list containing the \code{\link{top_ten_functions}} and \code{\link{last_ten_functions}}}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
