\name{ten_functions}
\alias{ten_functions}
\title{Top Ten and Last Ten Functions...}

\description{
  Top Ten and Last Ten Functions return a list that contains both the
  top ten functions and last ten functions
}
\value{list containing the \code{\link{top_ten_functions}} and \code{\link{last_ten_functions}}}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
