\name{top_ten_functions}
\alias{top_ten_functions}
\title{Top Ten Functions...}
\usage{top_ten_functions(fun_list=get_function_history())}
\description{Top Ten Functions
return the top ten unique functions that were executed}
\value{fuctions that are from a package and their help paths}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{fun_list}{function list}}
