\name{top_ten_functions}
\alias{top_ten_functions}
\title{List top 10 most frequently called functions.}
\usage{top_ten_functions(fun_list=get_function_history())}
\description{List top 10 most frequently called functions.}
\value{fuctions that are from a package and their help paths}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{\item{fun_list}{function list}}
