\name{topic_is_internal}
\alias{topic_is_internal}
\title{Internal topic function.}
\usage{
  topic_is_internal(help)
}
\arguments{
  \item{help}{\code{pkg_topic(}\emph{\code{topic}}\code{)}
  is checked to see if a keyword is "internal"}
}
\value{
  boolean
}
\description{
  Internal topic function.
}
\author{
  Barret Schloerke
}
\keyword{internal}

