\name{urlJSON_to_list}
\alias{urlJSON_to_list}
\title{URL made of JSON to list...}
\usage{urlJSON_to_list(url_string)}

\description{
  URL made of JSON to list
}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{internal}
\arguments{
  \item{url_string}{url that contains a JSON output to be turned into a list}
}
