% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-text.R
\name{write_text}
\alias{write_text}
\alias{read_text}
\title{Write to a Text File}
\usage{
write_text(path, lines, overwrite = FALSE)

read_text(path, n = -1L)
}
\arguments{
\item{path}{Path to target file. It is created if it does not exist, but the
parent directory must exist.}

\item{lines}{Character vector of lines. For \code{write_union()}, these are lines
to add to the target file, if not already present. For \code{write_over()},
these are the exact lines desired in the target file.}

\item{overwrite}{\code{logical(1)}. If file exists, over-write it?}

\item{n}{integer.  The (maximal) number of lines to
    read.  Negative values indicate that one should read up to the end of
    input on the connection.}
}
\description{
Writes as UTF-8 encoding and with \verb{\\n} line
endings on UNIX systems and \verb{\\r\\n} on Windows.
This avoids some of the unintended side-effects of
\code{usethis::write_over()} with R-projects.
}
\section{Functions}{
\itemize{
\item \code{read_text()}: a convenient wrapper to \code{\link[=readLines]{readLines()}} with default
UTF-8 encoding for reading text into an R session.

}}
